% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_abundance.R
\name{abundance}
\alias{abundance}
\alias{other}
\title{My title}
\usage{
abundance(censdata, type = "abund", alivecode = c("A"), mindbh = NULL,
  dbhunit = "mm", split1 = NULL, split2 = NULL)
}
\description{
My description

Mortality is the main function, and is constructed like
growth and recruitment. It requires two complete datasets, one per census,
with dbh, pom, and date for every individual of all species in at least 2 censuses (see Data Format).

Output of the mortality function is a list with components:
\itemize{
\item N, the number of individuals alive in the census 1 per category selected
\item D, the number of individuals no longer alive in census 2
\item rate, the mean annualized mortality rate constant per category selected, calculated as (log(N)-log(S))/time
\item upper, upper confidence limit of mean rate
\item lower, lower confidence limit of mean rate
\item time, mean time interval in years
\item date1, mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer)
\item date2, mean date in census 2
\item dbhmean, mean dbh in census 1 of individuals included
}

Pass the list to assemble.demography (in utilities.r) with type="m" to convert the list a data.frame.
}
\details{
My details
}
\section{My section}{

My cool section
}

\author{
R. C.
}
