% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{add_elevation_contrours}
\alias{add_elevation_contrours}
\title{Represent elevation with lines.}
\usage{
add_elevation_contrours(p, contour_size = 1, low = "blue",
  high = "red", bins = NULL)
}
\arguments{
\item{p}{A ggplot.}

\item{contour_size}{A number giving the size of the contour of elevation
lines. Passed to \code{ggplot2::stat_contour()} (see \code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}).}

\item{low, high}{A string giving a color of the elevation lines representing
low and high elevation.}

\item{bins}{A number giving the number of elevation lines to map.}
}
\description{
Represent elevation with lines.
}
\examples{
p <- plot_base_elevation(fgeo.data::luquillo_elevation)
add_elevation_contrours(p)
add_elevation_contrours(p, contour_size = 0.5, low = "grey", high = "black", bins = 4)
}
\seealso{
\code{\link[=plot_each_species]{plot_each_species()}}, \code{\link[=plot_species_or_elevation]{plot_species_or_elevation()}}, \code{\link[=plot_elevation]{plot_elevation()}},
\code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}.

Other functions to create or modify plot layers: \code{\link{add_elevation_labels}},
  \code{\link{add_species}}, \code{\link{axis_limits}},
  \code{\link{facet}}, \code{\link{hide}},
  \code{\link{plot_base_elevation}},
  \code{\link{theme_default}}
}
\concept{functions to create or modify plot layers}
