% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_limits}
\alias{add_limits}
\title{Add plot limits for quadrats 1-4.}
\usage{
add_limits(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A modified data frame.
}
\description{
Add plot limits for quadrats 1-4.
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_latest_tree_status}},
  \code{\link{add_limits_shrinked}},
  \code{\link{add_quadrat_and_subquadrat_from_list}},
  \code{\link{add_quadrat_to_df_list}},
  \code{\link{add_subquadrat}},
  \code{\link{identify_subquadrat}},
  \code{\link{prep_repulsive_tags}}
}
\keyword{internal}
