% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_quadrat_to_df_list}
\alias{add_quadrat_to_df_list}
\title{Add the name of each element of a list as a value in the variable quadrat}
\usage{
add_quadrat_to_df_list(df)
}
\arguments{
\item{df}{A list of data frames}
}
\value{
A list
}
\description{
Add the name of each element of a list as a value in the variable quadrat
}
\examples{
\dontrun{
df <- sin_q20[7:8]
df_list <- add_quadrat_to_df_list(df)
str(df_list)
# Demonstrate on only two quadrats
list_of_dfs <- sin_q20[c("109", "15")]
str(list_of_dfs)
add_quadrat_to_df_list(list_of_dfs)
}
}
\keyword{internal}
