% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapply_sp_elev.R
\name{add_sp}
\alias{add_sp}
\title{Represent species with points.}
\usage{
add_sp(p, data = NULL, fill = "sp", shape = 21, point_size = 3)
}
\description{
Represent species with points.
}
\seealso{
Other map components.: \code{\link{contour_elev}},
  \code{\link{fgeo_facets}},
  \code{\link{hide_axis_labels}},
  \code{\link{hide_legend_elev}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}}
}
