% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{add_species}
\alias{add_species}
\title{Represent species with points.}
\usage{
add_species(p, data = NULL, fill = "sp", shape = 21,
  point_size = 3)
}
\arguments{
\item{p}{A ggplot.}

\item{data}{A dataframe; passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{fill}{Character; either a colour or "sp", which maps each species to a
different color.}

\item{shape, point_size}{A number giving point size and shape (as in
\code{\link[graphics:points]{graphics::points()}}). Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
}
\description{
Represent species with points.
}
\examples{
# Small dataset with a few species for quick examples
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% some_sp)
p <- plot_base_census(census)

add_species(p)

add_species(p, fill = "white", shape = 22, point_size = 4)
}
\seealso{
\code{\link[=plot_species_or_elevation]{plot_species_or_elevation()}}, \code{\link[=plot_each_species]{plot_each_species()}},
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.

Other functions to create or modify plot layers: \code{\link{add_elevation_contrours}},
  \code{\link{add_elevation_labels}},
  \code{\link{axis_limits}}, \code{\link{facet}},
  \code{\link{hide}}, \code{\link{plot_base_elevation}},
  \code{\link{theme_default}}
}
\concept{functions to create or modify plot layers}
