% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tag.R
\name{add_subquad_limits}
\alias{add_subquad_limits}
\title{To a dataframe with \code{subquadrat} variable, add plot limits.}
\usage{
add_subquad_limits(df_with_subquad, quad_size = 20, shrink = 0.45)
}
\arguments{
\item{df_with_subquad}{A dataframe with the variable \code{subquadrat} that
defines the 1-4 subquadrats within each quadrat.}

\item{quad_size}{Size of each quadrat.}

\item{shrink}{A number, generally smaller than one, giving how much to
shrink the plot.}
}
\value{
A modified data frame.
}
\description{
This function helps fine tune the limits of your plots. You can live withouth
it, but this function gives you some extra control. The output is a modified
dataframe that includes the limits that, via \code{\link[=plot_repulsive_tags]{plot_repulsive_tags()}}, will be
passed to the arguments \code{xlim} and \code{ylim} of \code{\link[ggplot2:coord_fixed]{ggplot2::coord_fixed()}}.
}
\details{
Plots produced with \strong{ggplot} by default print with a margin around the
limits set by the user. To remove that extra margin and maximize space, this
function shrinks the limits a little. Be sure not to shrink so much that
you loose data; The numbers on the x and y axes should include the limits
that you expect. For example, if your plot is 20x20 meters, each subplot
will be 10x10 meters, so your plot axes should show either the 0 to 10, or
10 to 20. If you do not read 0 and/or 10, or 10 and/or 20, your plot may not
show the data you expect.
}
\examples{
\dontrun{
# not running because this example may be obsolete
library(dplyr)

# Showing only 1 quadrat to save space
with_subquad_list <- toy_list[1] \%>\%
  add_quadrat_and_subquadrat_from_list()
str(with_subquad_list)

# Pulling only one dataframe
with_subquad_df <- with_subquad_list[[1]][[1]]
head(with_subquad_df)

with_subquad_df \%>\%
  # The only "must be" is the variable `subquadrat`; we could remove `quadratname`
  select(-quadratname) \%>\%
  add_subquad_limits(quad_size = 20) \%>\%
  head()
}
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_quadrat_to_df_list}}
}
