% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-sp_elev.R
\name{autoplot.sp_elev}
\alias{autoplot.sp_elev}
\title{Quick plot of species distribution and topography.}
\usage{
\method{autoplot}{sp_elev}(object, ...)
}
\arguments{
\item{object}{An object created with \code{sp_elev()}.}

\item{...}{Other arguments passed to \code{plot_species_or_elevation()}.}
}
\value{
A "ggplot".
}
\description{
Automatically plot the \code{sp} and \code{elev} variables of a ForestGEO-like
dataset of class 'sp_elev'. You can create a 'sp_elev' \code{object} with:

\preformatted{object <- sp_elev(DATA-WITH-VARIABLE-sp, DATA-WITH-VARIABLE-elev)
}

See sections \strong{Usage} and \strong{Examples}.
}
\examples{
# Small dataset with a few species for quick examples
tree <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% c("PREMON", "CASARB"))
elevation_list <- fgeo.data::luquillo_elevation

autoplot(sp_elev(tree, elevation_list))

# Same
elevation_dataframe <- elevation_list$col
autoplot(sp_elev(tree, elevation_dataframe))

# Customize
p <- autoplot(sp_elev(tree, elevation_dataframe), fill = "red")
p
hide_color_legend(p)
}
\seealso{
\code{\link[=autoplot]{autoplot()}}, \code{\link[=sp_elev]{sp_elev()}}, \code{\link[=plot_species_or_elevation]{plot_species_or_elevation()}}.

Other autoplots: \code{\link{autoplot.elev}},
  \code{\link{autoplot.fgeo_habitat}},
  \code{\link{autoplot.sp}}
}
\concept{autoplots}
