% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_by_species.R
\name{autoplot_by_species}
\alias{autoplot_by_species}
\title{Generic function to create a list of autoplots by species.}
\usage{
autoplot_by_species(object, ...)
}
\arguments{
\item{object}{An object of supported S3 class.}

\item{...}{Other arguments passed to methods.}
}
\value{
A list of 'ggplots'.
}
\description{
Generic function to create a list of autoplots by species.
}
