% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_by_species.R
\name{autoplot_by_species.sp_elev}
\alias{autoplot_by_species.sp_elev}
\alias{autoplot_by_species.sp}
\title{List plots of species distribution and topography (good for pdf output).}
\usage{
\method{autoplot_by_species}{sp_elev}(object, species = "all",
  fill = "black", shape = 21, point_size = 3, contour_size = 0.5,
  low = "blue", high = "red", hide_color_legend = FALSE,
  bins = NULL, add_elevation_labels = TRUE, label_size = 3,
  label_color = "grey", xyjust = 1, fontface = "italic",
  xlim = NULL, ylim = NULL, custom_theme = NULL, ...)

\method{autoplot_by_species}{sp}(object, species = "all",
  fill = "black", shape = 21, point_size = 3,
  hide_color_legend = FALSE, xlim = NULL, ylim = NULL,
  custom_theme = NULL, ...)
}
\arguments{
\item{object}{An object created with \code{\link[=sp_elev]{sp_elev()}}.}

\item{species}{A character vector. Each element of the vector must be the
code for one species in the column \code{sp}. This function will produce as many
maps as elements in this vector. The string "all" (default) is a shortcut
to map all unique codes in the column \code{sp}.}

\item{fill}{Character; either a colour or "sp", which maps each species to a
different color.}

\item{shape}{A number giving point size and shape (as in
\code{\link[graphics:points]{graphics::points()}}). Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{point_size}{A number giving point size and shape (as in
\code{\link[graphics:points]{graphics::points()}}). Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{contour_size}{A number giving the size of the contour of elevation
lines. Passed to \code{ggplot2::stat_contour()} (see \code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}).}

\item{low}{A string giving a color of the elevation lines representing
low and high elevation.}

\item{high}{A string giving a color of the elevation lines representing
low and high elevation.}

\item{hide_color_legend}{Logical; \code{TRUE} hides the color legend.}

\item{bins}{A number giving the number of elevation lines to map.}

\item{add_elevation_labels}{Logical; \code{FALSE} hides elevation labels.}

\item{label_size}{A number (\code{label_size}) or character
string (\code{label_color} and \code{fontface}) giving the size, colour and fontface
of the text labels for the elevation lines.}

\item{label_color}{A number (\code{label_size}) or character
string (\code{label_color} and \code{fontface}) giving the size, colour and fontface
of the text labels for the elevation lines.}

\item{xyjust}{A number to adjust the position of the text labels of the
elevation lines.}

\item{fontface}{A number (\code{label_size}) or character
string (\code{label_color} and \code{fontface}) giving the size, colour and fontface
of the text labels for the elevation lines.}

\item{xlim}{A vector of length 2, for example \code{c(0, 500)}, giving the
minimum and maximum limits of the vertical and horizontal coordinates.}

\item{ylim}{A vector of length 2, for example \code{c(0, 500)}, giving the
minimum and maximum limits of the vertical and horizontal coordinates.}

\item{custom_theme}{A valid \code{\link[ggplot2:theme]{ggplot2::theme()}}. \code{NULL} uses the default
theme \code{\link[=theme_default]{theme_default()}}.}

\item{...}{Other arguments passed to methods.}
}
\description{
These functions extend the functionality of \code{\link[=autoplot.sp]{autoplot.sp()}} and
\code{\link[=autoplot.elev]{autoplot.elev()}} and return not a single plot but a list of plots. It is
ideal to print a .pdf file with one plot per page. They automatically plot
the variables \code{sp} and \code{elev} of a ForestGEO-like dataset of class 'sp_elev'.
\itemize{
\item Create a 'sp' \code{object} with:
}

\preformatted{    object <- sp(DATA-WITH-VARIABLE-sp)
}
\itemize{
\item Create a 'sp_elev' \code{object} with:
}

\preformatted{    object <- sp_elev(DATA-WITH-VARIABLE-sp, DATA-WITH-VARIABLE-elev)
}

See sections \strong{Usage} and \strong{Examples}.
}
\details{
\code{autoplot_by_species(sp_elev(DATA-WITH-VARIABLE-sp)} (without elevation data)
is equivalent to \code{autoplot_by_species(sp(DATA-WITH-VARIABLE-sp))}.

\strong{fgeo.map} wraps the \strong{ggplot2} package. For more control you
can use \strong{ggplot2} directly.
}
\examples{
# Species ---------------------------------------------------------------
# Small dataset with a few species for quick examples
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.x::tree5, sp \%in\% some_sp)

plots_list <- autoplot_by_species(sp(census))
# Show the first plot in the list
plots_list[[1]]

\dontrun{
# Print all plots in the list
pdf("map.pdf", paper = "letter", height = 10.5, width = 8)
plots_list
dev.off()
}

# Species and or elevation ----------------------------------------------
# Species only: Same as autoplot_by_species(sp(census))
autoplot_by_species(sp_elev(census))

# Species and elevation
census_elevation <- sp_elev(census, fgeo.x::elevation)
p <- autoplot_by_species(census_elevation)
# Showing only the first plot
p[[1]]
}
\seealso{
\code{\link[=autoplot]{autoplot()}}, \code{\link[=sp]{sp()}}, \code{\link[=sp_elev]{sp_elev()}}.

Other functions to plot elevation: \code{\link{autoplot.sp_elev}},
  \code{\link{elev}}, \code{\link{sp_elev}}

Other functions to plot species: \code{\link{autoplot.sp_elev}},
  \code{\link{sp_elev}}, \code{\link{sp}}
}
\concept{functions to list plots from different ForestGEO classes}
\concept{functions to plot elevation}
\concept{functions to plot species}
