% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci12vft_mini}
\alias{bci12vft_mini}
\title{An example ViewFullTable data set.}
\format{A tibble with 4,374 rows and 28 variables.}
\source{
https://doi.org/10.5479/data.bci.20130603.
}
\usage{
bci12vft_mini
}
\description{
ViewFullTable with data of a only 10 quadrats of the censuses 6 and 7 of the
50 hectare plot of Barro Colorado Island downloaded from:
https://doi.org/10.5479/data.bci.20130603. Explanation by Suzanne Lao
https://goo.gl/fg2nqh.
}
\keyword{datasets}
