% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bciYYsNmini}
\alias{bciYYsNmini}
\alias{bci12s1mini}
\alias{bci12s2mini}
\alias{bci12s3mini}
\alias{bci12s4mini}
\alias{bci12s5mini}
\alias{bci12s6mini}
\alias{bci12s7mini}
\title{Subset of stem data from Barro Colorado Island, Panama.}
\format{A dataframe with 2165 rows and 20 variables.
Variables are defined at: http://ctfs.si.edu/Public/DataDict/data_dict.php.}
\source{
From \url{https://repository.si.edu/handle/10088/20925}, file
bci.stem#.rda.
}
\description{
Stem data corresponding to a subset of 1000 randomly selected trees from
Barro Colorado Island, Panama (data released in 2012). For every bciYYsNmini,
"YY" and "N" indicate the year and census number. For example, bci12s1mini
contains BCI data released in 2012 from census 1.
}
