% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_crucial_names}
\alias{check_crucial_names}
\title{Check if an object contains specific names.}
\usage{
check_crucial_names(x, nms)
}
\arguments{
\item{x}{A named object.}

\item{nms}{String; names expected to be found in \code{x}.}
}
\value{
Invisible \code{x}, or an error with informative message.
}
\description{
Check if an object contains specific names.
}
\examples{
v <- c(x = 1)
check_crucial_names(v, "x")

dfm <- data.frame(x = 1)
check_crucial_names(dfm, "x")
}
\seealso{
Other Functions to check inputs.: \code{\link{check_unique}}

Other Functions for developers.: \code{\link{check_unique}},
  \code{\link{exists_in_pkg}}
}
