% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_unique}
\alias{check_unique}
\alias{check_unique_vector}
\title{Report if a vector or a variable of a dataframe is duplicated.}
\usage{
check_unique(x, x_var, cond = "warning", msg = NULL)

check_unique_vector(v, cond, msg = NULL)
}
\arguments{
\item{x}{A dataframe.}

\item{x_var}{String; the name of a variable of \code{x}.}

\item{cond}{String; the name of a function that outputs a condition: one of
"warning", "stop", "message".}

\item{msg}{String; a custom message.}

\item{v}{A vector.}
}
\value{
Invisible \code{v} or a condition and a message.
}
\description{
Report if a vector or a variable of a dataframe is duplicated.
}
\examples{
# On a vector
unique_v <- rep(1, 3)
num <- c(1:3)
chr <- c(letters[1:3])
check_unique_vector(unique_v, "warning")
check_unique_vector(num, "warning")
check_unique_vector(chr, "message", "Do something")

# On a dataframe
.df <- data.frame(a = 1:3, b = 1, stringsAsFactors = FALSE)

check_unique(.df, "a")
check_unique(.df, "a", "message", "do this")
# Silent
check_unique(.df, "b", "warning", "do this")
}
\seealso{
Other Functions to check inputs.: \code{\link{check_crucial_names}}

Other Functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{exists_in_pkg}}
}
