% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codify_explicitely.R
\name{codify_explicitely}
\alias{codify_explicitely}
\title{Add alternatives to symbol, that are easier to understand.}
\usage{
codify_explicitely(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
The modified data frame.
}
\description{
Add alternatives to symbol, that are easier to understand.
}
\examples{
sin_q20 <- map(sin_q20, codify_explicitely)
}
