% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{contour_elev}
\alias{contour_elev}
\title{Represent elevation with lines.}
\usage{
contour_elev(p, contour_size = 1, low = "blue", high = "red",
  bins = NULL)
}
\arguments{
\item{p}{A ggplot.}

\item{contour_size}{A number giving the size of the contour of elevation
lines. Passed to \code{\link[ggplot2:stat_contour]{ggplot2::stat_contour()}}.}

\item{low, high}{A string giving a color of the elevation lines representing
low and high elevation.}

\item{bins}{A number giving the number of elevation lines to map.}
}
\description{
Represent elevation with lines.
}
\examples{
elevation <- fgeo.data::luquillo_elevation

base_elev <- map_gx_gy_elev(elevation)

contour_elev(base_elev)

contour_elev(
  base_elev, contour_size = 0.5, low = "grey", high = "black", bins = 4
)
}
\seealso{
\code{\link[=maply_sp_elev]{maply_sp_elev()}}, \code{\link[=map_sp_elev]{map_sp_elev()}}, \code{\link[=map_elev]{map_elev()}},
\code{\link[ggplot2:stat_contour]{ggplot2::stat_contour()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{hide}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
\concept{map components.}
