% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{one_quadrat}
\alias{four_quadrats}
\title{Data sets for examples; adapted from \link[bciex:bci12vft_mini]{bciex::bci12vft_mini}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 143 rows and 28 columns.}
\usage{
one_quadrat

four_quadrats
}
\description{
Data sets for examples; adapted from \link[bciex:bci12vft_mini]{bciex::bci12vft_mini}.
}
\examples{
library(tibble)
one_quadrat
four_quadrats
}
\keyword{datasets}
