% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{facets}
\alias{facets}
\alias{facet_wrap_sp}
\alias{facet_h_sp}
\alias{facet_v_sp}
\title{Facet maps vertically, hirizontally, or wrap them to fit a page.}
\usage{
facet_wrap_sp(p, ...)

facet_h_sp(p, ...)

facet_v_sp(p, ...)
}
\arguments{
\item{p}{A ggplot.}

\item{...}{Arguments passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.}
}
\description{
Facet maps vertically, hirizontally, or wrap them to fit a page.
}
\seealso{
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}},
  \code{\link{hide_axis_labels}},
  \code{\link{hide_legend_elev}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}}
}
