% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.map-package.R
\docType{package}
\name{fgeo.map-package}
\alias{fgeo.map}
\alias{fgeo.map-package}
\title{fgeo.map: Map Species, Trees and Topography}
\description{
Map species, trees and topography.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.map}
  \item \url{https://forestgeo.github.io/fgeo.map/}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.map/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
