% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.plot-package.R
\docType{package}
\name{fgeo.plot-package}
\alias{fgeo.plot}
\alias{fgeo.plot-package}
\title{fgeo.plot: Plot ForestGEO Data}
\description{
To help you access, transform, analyze, and
    visualize ForestGEO data, we developed a collection of R packages
    (<https://forestgeo.github.io/fgeo/>). This package, in particular,
    helps you to plot ForestGEO data. To learn more about ForestGEO visit
    <http://www.forestgeo.si.edu/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.plot}
  \item \url{https://forestgeo.github.io/fgeo.plot/}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.plot/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (0000-0002-1986-7988) [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
