% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo}
\alias{fgeo}
\title{Clasify ForestGEO-like data or return unknown.}
\usage{
fgeo(.data)
}
\arguments{
\item{.data}{A list or a dataframe.}
}
\value{
A string.
}
\description{
Clasify ForestGEO-like data or return unknown.
}
\examples{
fgeo(fgeo.data::luquillo_elevation)
fgeo(fgeo.data::luquillo_elevation$col)
fgeo(fgeo.data::luquillo_stem6_random)
fgeo(fgeo.data::luquillo_tree6_random)
fgeo(list(1))
fgeo(data.frame(x = 1))
}
