% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{fgeo_facets}
\alias{fgeo_facets}
\alias{facet_wrap_sp}
\alias{facet_grid_sp_h}
\alias{facet_grid_sp_v}
\title{Facets to add on top of ggplots based on ForestGEO's data.}
\usage{
facet_wrap_sp(...)

facet_grid_sp_h(...)

facet_grid_sp_v(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.}
}
\description{
Facets to add on top of ggplots based on ForestGEO's data.
}
\seealso{
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.
}
