% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{fgeo_geoms}
\alias{fgeo_geoms}
\alias{geom_point_sp}
\title{Geoms to add on top of ggplots based on ForestGEO's data.}
\usage{
geom_point_sp(data = NULL, drop_fill = FALSE, shape = 21, ...)
}
\description{
Geoms to add on top of ggplots based on ForestGEO's data.
}
\seealso{
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
}
