% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{four_quadrats}
\alias{four_quadrats}
\title{Data set for examples; adapted from \link[bciex:bci12vft_mini]{bciex::bci12vft_mini}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 178 rows and 30 columns.}
\usage{
four_quadrats
}
\description{
Data set for examples; adapted from \link[bciex:bci12vft_mini]{bciex::bci12vft_mini}.
}
\keyword{datasets}
