% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak_plot.R
\name{get_settings}
\alias{get_settings}
\alias{get_subtitle}
\alias{get_subtitle_line1}
\alias{get_subtitle_line2}
\alias{get_subtitle_line3}
\alias{get_site_name}
\alias{get_size_plot_title}
\alias{get_size_plot_subtitle}
\alias{get_size_axis_text}
\alias{get_size_tag}
\alias{get_size_point}
\alias{get_shape_point}
\alias{get_panel_grid_major_colour}
\alias{get_panel_grid_minor_colour}
\alias{get_panel_grid_minor_linetype}
\alias{get_panel_background_fill}
\alias{get_legend_position}
\alias{get_legend_title}
\alias{get_axis_ticks}
\title{Functions to change plot layout.}
\usage{
get_subtitle(line1 = get_subtitle_line1(), line2 = get_subtitle_line2(),
  line3 = get_subtitle_line3())

get_subtitle_line1()

get_subtitle_line2()

get_subtitle_line3()

get_site_name(x = "Sinharaja 2017")

get_size_plot_title(x = 20)

get_size_plot_subtitle(x = 12)

get_size_axis_text(x = 12)

get_size_tag(x = 3)

get_size_point(x = 1.5)

get_shape_point(x = c(19, 4))

get_panel_grid_major_colour(x = "grey")

get_panel_grid_minor_colour(x = "black")

get_panel_grid_minor_linetype(x = "dotted")

get_panel_background_fill(x = "white")

get_legend_position(x = "top")

get_legend_title(x = element_blank())

get_axis_ticks(x = element_blank())
}
\arguments{
\item{line1, line2, line3}{String. First, second and third subtitle lines.}

\item{x}{A dafault setting.}
}
\description{
Functions to change plot layout.
}
\seealso{
Other functions to change plot layout.: \code{\link{get_theme}}
}
