% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_subtitle}
\alias{get_subtitle}
\title{Get settings for plot_repel().}
\usage{
get_subtitle(site_string = "(Sinharaja)")
}
\arguments{
\item{site_string}{A string giving the name of the site. E.g. "Sinharaja".}
}
\description{
Get settings for plot_repel().
}
\keyword{internal}
