% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_theme}
\alias{get_theme}
\title{A theme that is customized for plot_repulsive_tags().}
\usage{
get_theme(panel_grid_major_colour = "grey",
  panel_grid_minor_colour = "black", panel_grid_minor_linetype = "dotted",
  panel_background_fill = "white", plot_title_size = 20,
  plot_subtitle_size = 12, legend_position = "top",
  axis_ticks = element_blank(), axis_text_size = 12,
  legend_title = element_blank(), ...)
}
\arguments{
\item{panel_grid_major_colour, panel_grid_minor_colour, panel_grid_minor_linetype, panel_background_fill, plot_title_size, plot_subtitle_size, legend_position, axis_ticks, axis_text_size, legend_title, ...}{Arguments passedto
\code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A customized \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
Theme customized for \code{\link[=plot_repel]{plot_repel()}}. You further customize passing more
arguments to \code{\link[ggplot2:theme]{ggplot2::theme()}} via \code{...} or replace this theme by a
completely new \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\examples{
ggplot2::theme()
get_theme()
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
