% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{header_dbh_bubbles}
\alias{header_dbh_bubbles}
\title{Pre-made headers.}
\usage{
header_dbh_bubbles(lang = "english")
}
\arguments{
\item{lang}{String; Language of the header: Either "english" or "spanish".}
}
\description{
Pre-made headers.
}
\examples{
cat(header_dbh_bubbles())
}
\seealso{
Other functions to plot dbh bubbles: 
\code{\link{plot_dbh_bubbles_by_quadrat}()},
\code{\link{theme_dbh_bubbles}()}
}
\concept{functions to plot dbh bubbles}
\keyword{internal}
