% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{header_tag_status}
\alias{header_tag_status}
\title{Pre-made headers.}
\usage{
header_tag_status()
}
\description{
Pre-made headers.
}
\examples{
cat(header_tag_status())
}
\seealso{
Other functions to plot tag status: 
\code{\link{plot_tag_status_by_subquadrat}()},
\code{\link{theme_tag_status}()}
}
\concept{functions to plot tag status}
\keyword{internal}
