% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fun_settings.R
\name{headers}
\alias{headers}
\alias{header_map_tag}
\alias{header_map_quad}
\title{Pre-made headers.}
\usage{
header_map_tag()

header_map_quad()
}
\description{
Pre-made headers.
}
\examples{
header_map_tag()
header_map_quad()
}
