% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{hide}
\alias{hide}
\alias{hide_axis_labels}
\alias{hide_legend_color}
\title{Hide elements of a plot.}
\usage{
hide_axis_labels(p)

hide_legend_color(p)
}
\arguments{
\item{p}{A ggplot.}
}
\description{
\code{hide_axis_labels} hides axis labels.
\code{hide_legend_color} hides the color legend, for example, of elevation lines.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
