% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{hide_axis_labels}
\alias{hide_axis_labels}
\title{Hide axis labels.}
\usage{
hide_axis_labels(p)
}
\arguments{
\item{p}{A ggplot.}
}
\description{
Hide axis labels.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{facets}},
  \code{\link{hide_legend_elev}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}}
}
