% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapply_sp_elev.R
\name{hide_legend_elev}
\alias{hide_legend_elev}
\title{Hide the color legend of elevation lines.}
\usage{
hide_legend_elev(p)
}
\description{
Hide the color legend of elevation lines.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{fgeo_facets}},
  \code{\link{hide_axis_labels}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}}
}
