% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapply_sp_elev.R
\name{label_elev}
\alias{label_elev}
\title{Label elevation lines.}
\usage{
label_elev(p, label_size = 3, label_color = "grey", xyjust = 1,
  fontface = "italic")
}
\description{
Label elevation lines.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{fgeo_facets}},
  \code{\link{hide_axis_labels}},
  \code{\link{hide_legend_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}}
}
