% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_plot_repel.R
\name{lapply_plot_repel}
\alias{lapply_plot_repel}
\alias{plot_repel}
\title{Plot trees in subquadrats, avoiding tree tags to overlap.}
\usage{
lapply_plot_repel(df_list, point_size = 1.5, id_size = 3, text_size = 15)

plot_repel(df, point_size, id_size, text_size)
}
\arguments{
\item{df_list}{A list of data frames.}

\item{point_size}{Point size passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{id_size}{Size of tree tags, passed to \code{\link[=geom_text_repel]{geom_text_repel()}}.}

\item{text_size}{Text size passed to \code{\link[ggplot2:element_text]{ggplot2::element_text()}}.}

\item{df}{A data frame as wranglec by \code{lapply_plot_repel()}.}

\item{...}{Arguments passed to \code{plot_repel()}.}
}
\value{
A list of plots that can be wrapped around pdf(onefile = TRUE).
}
\description{
Plot trees in subquadrats, avoiding tree tags to overlap.
}
