% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_map_sp.R
\name{lapply_plot_sp}
\alias{lapply_plot_sp}
\title{Plot species from a list or vector.}
\usage{
lapply_plot_sp(species, cns_data, ...)
}
\arguments{
\item{species}{A string of the species codes to plot (\code{sp}).}

\item{cns_data}{Census data.}

\item{...}{Arguments passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
}
\value{
A list of plots. Usually you will want to save the list on a file (
e.g. a .pdf, .png, etc.)
}
\description{
Plot species from a list or vector.
}
\section{Acknowledgements}{

Thanks to Gabriel Arellano and "Kenfack, David" for ideas and feedback.
}

\examples{
\dontrun{
# Ensure you have these packages
library(ggplot2)
library(dplyr)

# Minimal data to show
library(bci)
cns_data <- bci::bci12full7
# Filtering only 3 species for a minimal example.
(species_selected <- unique(cns_data$sp)[1:3])
cns_data_sub <- filter(cns_data, sp \%in\% species_selected)
# Sampling only 1000 rows for a quick example
cns_data_sub <- sample_n(cns_data_sub, 1000)

# Selecting all species in the example dataset
all_species <- unique(cns_data_sub$sp)
pdf(onefile = TRUE)
lapply_plot_sp(species = all_species, cns_data = cns_data_sub)
dev.off()

pdf(onefile = TRUE)
lapply_plot_sp(
  species = all_species, cns_data = cns_data_sub,
  # Change the look of the points  passed to ?geom_point
  color = "blue", size = 3, shape = 1
)
dev.off()

}
}
\seealso{
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}, \code{\link[grDevices:pdf]{grDevices::pdf()}}, \code{\link[grDevices:png]{grDevices::png()}}.
}
