% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{limit_gx_gy}
\alias{limit_gx_gy}
\title{Set the map limits.}
\usage{
limit_gx_gy(p, xlim = NULL, ylim = NULL)
}
\arguments{
\item{p}{A ggplot.}

\item{xlim, ylim}{A vector of length 2, for example \code{c(0, 500)}, giving the
minimum and maximum limits of the vertical and horizontal coordinates.}
}
\description{
Set the map limits.
}
\examples{
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_stem5_random, sp \%in\% some_sp)
p <- map_gx_gy(census)
p

limit_gx_gy(p, xlim = c(0, 100), ylim = c(0, 400))

fgeo.data::luquillo_elevation \%>\% 
  map_gx_gy_elev() \%>\% 
  limit_gx_gy(xlim = c(0, 100), ylim = c(0, 400))
}
\seealso{
\code{\link[=plot_species_or_elevation]{plot_species_or_elevation()}}, \code{\link[=plot_each_species]{plot_each_species()}},
\code{\link[ggplot2:coord_fixed]{ggplot2::coord_fixed()}}.

Other functions to create or modify plot layers: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{hide}},
  \code{\link{label_elev}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
\concept{functions to create or modify plot layers}
