% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-package.R
\docType{package}
\name{map-package}
\alias{map}
\alias{map-package}
\title{map: Map ForestGEO's Data}
\description{
Map ForestGEO's data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/map}
  \item \url{https://forestgeo.github.io/map/}
  \item Report bugs at \url{https://github.com/forestgeo/map/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
