% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{map_gx_gy_elev}
\alias{map_gx_gy_elev}
\alias{map_gx_gy_elev}
\alias{map_gx_gy}
\title{Map a base over which other map components can later be added.}
\usage{
map_gx_gy_elev(data)

map_gx_gy(data)
}
\arguments{
\item{data}{A dataframe; passed to [ggplot2::ggplot()].}
}
\description{
Map a base over which other map components can later be added.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{facets}},
  \code{\link{hide_axis_labels}},
  \code{\link{hide_legend_elev}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}
}
