% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tag.R
\name{map_tag}
\alias{map_tag}
\title{Map tree tags}
\usage{
map_tag(vft, dim_x = 20, div_x = 5, dim_y = dim_x, div_y = div_x,
  site_name = "Site Name, YYYY", point_shape = c(19, 4), point_size = 1.5,
  tag_size = 3, header = get_header(), theme = get_theme())
}
\arguments{
\item{vft}{A ViewFullTable.}
}
\description{
Map tree tags
}
\examples{
\dontrun{
}
}
