% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_edgy_dead.R
\name{nms_has_any}
\alias{nms_has_any}
\title{Functions to detect and extract names.}
\usage{
nms_has_any(x, ...)
}
\arguments{
\item{x}{A named object.}

\item{...}{Strings; each of the names that need to be checked.}
}
\description{
These functions are handy to work with fgeo's data structures because the
same variablel may be named differently in different data sets. For example,
the variable status is called \code{Status} or \code{status} in viewfull or census
(tree and stem) tables.
}
\details{
\code{nms_has_any()}: Checks if an object has any of the provided names.
\itemize{
\item Returns a logical value.
\code{nms_detect()}: Checks if an object has the provided names.
\item Returns a logical vector.
\code{nms_extract_all()}: Finds the names that match the provided names.
\item Returns a character vector.
\code{nms_extract1()}: Finds the first name that matchs the provided names.
\item Returns a character string.
}
}
\keyword{internal}
