% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{pad}
\alias{pad}
\title{Right-pad strings.}
\usage{
pad(string, total_width = 55, side = "right", pad = "_")
}
\arguments{
\item{total_width}{Total width of the string after pasting each string in the
vector of strings.}

\item{pad, string, side, }{... Arguments passed to \code{\link[stringr:str_pad]{stringr::str_pad()}}.}
}
\value{
A string which total number of characters is \code{total_width}.
}
\description{
Right-pad strings.
}
\examples{
pad(c("Hello", "world"))
pad(c("Hello", "world"), total_width = 70, pad = ".")
pad(c("Hello", "world"), total_width = 70, pad = ".", side = "left")
}
