% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tag.R
\name{paginate}
\alias{paginate}
\title{Label the four divisions of a quadrat -- each with four subquadrats.}
\usage{
paginate(x, bl = 1, br = 2, tr = 3, tl = 4)
}
\arguments{
\item{x}{A dataframe with the variable \code{subquadrat}.}

\item{bl, br, tr, tl}{Number or character giving the label of the four
subquadrats on each or the four divisions of a quadrat: bottom left (bl),
bottom right (br), top right (tr), and top left (tl).\preformatted{The four divisions of a quadrat (bl, br, tr, tl), each with four subquadrats:

tl       tr
----     -----
14 24    34 44
13 23    33 43

bl       br
----     -----
12 22    32 42
11 21    31 41
}}
}
\value{
A modified version of the input with the additional variable \code{page}.
}
\description{
This function makes it possible for \code{\link[=map_tag]{map_tag()}} to plot each individual map.
Each map corresponds to one page and includes four subquadrats. There are a
total of four maps per quadrat (i.e. a total of 16 subquadrats per quadrat).
}
\examples{
library(dplyr)
viewfulltable <- tribble(
   ~QX,  ~QY,
  17.9,    0,
   4.1,   15,
   6.1, 17.3,
   3.8,  5.9,
   4.5, 12.4,
   4.9,  9.3,
   9.8,  3.2,
  18.6,  1.1,
  17.3,  4.1,
   1.5, 16.3
)
with_subquad <- fgeo.utils::add_subquad(viewfulltable, 20, 20, 5, 5)
paginate(with_subquad)
paginate(with_subquad, "a", "b", "c", "d")
paginate(with_subquad, "bottom left", "bottom right", "top left", "top right")
}
\seealso{
\code{\link[fgeo.utils:add_subquad]{fgeo.utils::add_subquad()}}.
}
