% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_vft.R
\name{pick_vft}
\alias{pick_vft}
\title{Pick ViewFullTables to create small example datasets.}
\usage{
pick_vft(.data, n_censuses = NULL, n_quadrats = NULL, n_rows = NULL)
}
\arguments{
\item{.data}{A ViewFullTable}

\item{n_censuses, n_quadrats}{Number of censuses and quadrats to pick.}

\item{n_rows}{Number of rows to sample after picking \code{n_censuses} and
\code{n_quadrats}.}
}
\value{
A tibble (dataframe).
}
\description{
This function helps you to create small datasets for examples. It defaults to
returning the input ViewFullTable "as is", but provides arguments to pick a
specific number of censuses, quadrats and rows.
}
\examples{
# Same
dim(fgeo.x::vft_4quad)
dim(pick_vft(fgeo.x::vft_4quad))

dim(
  pick_vft(
    fgeo.x::vft_4quad, 
    n_censuses = 2, 
    n_quadrats = 2, 
    n_rows = 50
 )
)

}
\concept{functions for fgeo vft}
\concept{functions to pick or drop rows of a ForestGEO dataframe}
