% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fgeo_habitat.R
\name{plot.fgeo_habitat}
\alias{plot.fgeo_habitat}
\title{Plot objects of class "fgeo_habitat".}
\usage{
\method{plot}{fgeo_habitat}(x, ...)
}
\arguments{
\item{x}{An S3 object of class "fgeo_habitat".}

\item{...}{Other arguments passed to methods.}
}
\value{
An object of class "ggplot".
}
\description{
Plot objects of class "fgeo_habitat".
}
\examples{
\dontrun{
library(fgeo.tool)
elev_list <- fgeo.data::luquillo_elevation
habitats <- fgeo_habitat(elev_list, gridsize = 20, n = 4)
plot(habitats)
}
}
\concept{Methods for common generics}
