% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tt_df.R
\name{plot.tt_df}
\alias{plot.tt_df}
\title{Plot objects of class "tt_df".}
\usage{
\method{plot}{tt_df}(x, ...)
}
\arguments{
\item{x}{An object of class "tt_df".}

\item{...}{Other arguments passed to methods.}
}
\value{
A ggplot object.
}
\description{
Plot objects of class "tt_df".
}
\examples{
\dontrun{
library(fgeo.tool)
library(fgeo.habitat)
census <- filter(luquillo_top3_sp, status == "A", dbh >= 10)
habitat <- luquillo_habitat
species <- c("CASARB", "PREMON", "SLOBER")

tt_df <- to_df(tt_test(census, species, habitat))
plot(tt_df)
}
}
