% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{plot_base_elevation}
\alias{plot_base_elevation}
\alias{plot_base_census}
\title{Map a base over which other map components can later be added.}
\usage{
plot_base_elevation(data)

plot_base_census(data)
}
\arguments{
\item{data}{A dataframe; passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}
}
\description{
Map a base over which other map components can later be added.
}
\examples{
# Small dataset with a few species for quick examples
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% some_sp)
elevation <- fgeo.data::luquillo_elevation

# These functions look alike but internally they are not.
base_census <- plot_base_census(census)
base_census
base_elevation <- plot_base_elevation(elevation)
base_elevation

# Use plot_base_census() as a base for layers that need census data
add_species(base_census)
# This fails because the base plot has no information about elevation
\dontrun{
  add_elevation_contrours(base_census)
}

# Use plot_base_elevation() as a base for layers that need elevation data
add_elevation_contrours(base_elevation)
# This fails because the base plot has no information about species
\dontrun{
  add_species(base_elevation)
}
}
\seealso{
\code{\link[=plot_species_or_elevation]{plot_species_or_elevation()}}, \code{\link[=plot_elevation]{plot_elevation()}},
\code{\link[=plot_each_species]{plot_each_species()}}, \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.

Other functions to create or modify plot layers: \code{\link{add_elevation_contrours}},
  \code{\link{add_elevation_labels}},
  \code{\link{add_species}}, \code{\link{axis_limits}},
  \code{\link{facet}}, \code{\link{hide}},
  \code{\link{theme_default}}
}
\concept{functions to create or modify plot layers}
