% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_repulsive_tags}
\alias{plot_repulsive_tags}
\alias{lapply_plot_repulsive_tags}
\alias{plot_repulsive_tags}
\title{Plot trees in subquadrats, avoiding tree tags to overlap.}
\usage{
lapply_plot_repulsive_tags(prep_df_list, site_name = site_name,
  point_shape = c(19, 4), point_size = 1.5, tag_size = 3,
  header = get_header(), theme = get_theme())

plot_repulsive_tags(prep_df, site_name, point_shape, point_size, tag_size,
  header, theme)
}
\arguments{
\item{prep_df_list}{A list of data frames, each prapared specifically for
\link{plot_repulsive_tags}.}

\item{site_name}{A sting of the site name for the plot title.}

\item{point_shape}{Point shape. Passed to \code{value} in
\code{\link[ggplot2:scale_shape_manual]{ggplot2::scale_shape_manual()}}. See \code{pch} values in \code{\link[graphics:points]{graphics::points()}}.}

\item{point_size}{Point size. Passed to \code{size} in \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{tag_size}{Tag size. Passed to \code{size} in \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}.}

\item{header}{A string giving the plot header. An easy way to make a
three-line header is with \code{\link[=get_header]{get_header()}}.}

\item{theme}{A \code{\link[ggplot2:theme]{ggplot2::theme()}}. A theme customized for
\code{\link[=plot_repulsive_tags]{plot_repulsive_tags()}} is available via \code{\link[=get_theme]{get_theme()}}, which you can
further customize or you can create a complete new theme with
\code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{prep_df}{A data frames prapared specifically for \link{plot_repulsive_tags}.}
}
\value{
A list of plots that can be wrapped around pdf(onefile = TRUE).
}
\description{
Plot trees in subquadrats, avoiding tree tags to overlap.
}
\examples{
# The sinharaja package is local and the data set sinh_q20 is private.
list_of_dataframes <- sinharaja::sinh_q20[1:2]
prepared <- prep_repulsive_tags(list_of_dataframes)
plot_list <- lapply_plot_repulsive_tags(prepared, site_name = "Sinharaja")
\dontrun{
plot_list[[1]]
}

# Print each plot on a plage of a single .pdf file
\dontrun{
pdf(onefile = TRUE, paper = "a4", width = 11, height = 11)
plot_list
dev.off()
}
}
