% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{plot_species_or_elevation}
\alias{plot_species_or_elevation}
\alias{plot_elevation}
\title{Map species and elevation data.}
\usage{
plot_species_or_elevation(census, elevation = NULL, fill = "black",
  shape = 21, wrap = TRUE, point_size = 3, contour_size = 0.5,
  low = "blue", high = "red", hide_legend_color = FALSE,
  bins = NULL, label_elev = TRUE, label_size = 3,
  label_color = "grey", xyjust = 1, fontface = "italic",
  xlim = NULL, ylim = NULL, custom_theme = NULL)

plot_elevation(elevation, contour_size = 0.5, low = "blue",
  high = "red", hide_legend_color = FALSE, bins = NULL,
  label_elev = TRUE, label_size = 3, label_color = "grey",
  xyjust = 1, fontface = "italic", xlim = NULL, ylim = NULL,
  custom_theme = NULL)
}
\value{
\itemize{
\item \code{plot_each_species()} returns a list of ggplots.
\item \code{plot_elevation()} and \code{plot_species_or_elevation()} return a ggplot.
}
}
\description{
These functions wrap a number of map elements for convenience:
\itemize{
\item Use \code{plot_species_or_elevation()} to map species and optionally elevation
data in a single page. You can map multiple species on the same plot or you
can facet the output to map each species on a single plot and all plots in a
single page.
\item Use \code{plot_elevation()} if you want to map only elevation in the simplest
way.
\item Use \code{plot_each_species()} to apply the function
\code{plot_species_or_elevation()} to each species in a census dataset. The output
is not a map but a list of maps, one per species, that can be printed on a
.pdf file.
}
}
\examples{
# Small dataset with a few species for quick examples
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% some_sp)
elevation <- fgeo.data::luquillo_elevation

# Simplest way to map elevation data only
plot_elevation(elevation)

plot_species_or_elevation(census)

plot_species_or_elevation(census, elevation)

# For maximum control, you can compose maps as you like
# Traditional sintax: g(f(x))
contour_elev(map_gx_gy_elev(elevation))

# With the pipe: f(x) \%>\% g()
map_gx_gy_elev(elevation) \%>\%
  contour_elev()

# The traditional sintax is hard to read when you compose multiple functions.
# With the pipe, readability isn't affected by the number of functions.
map_gx_gy_elev(elevation) \%>\%
  limit_gx_gy(xlim = c(-100, 400)) \%>\%
  contour_elev(contour_size = 0.5) \%>\%
  label_elev(label_color = "red") \%>\%
  hide_axis_labels() \%>\%
  hide_legend_color() \%>\%
  add_sp(census, point_size = 5) \%>\%
  wrap("sp") \%>\%
  theme_default(legend.position = "top")
}
\concept{functions to create a single plot}
