% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prep_repulsive_tags}
\alias{prep_repulsive_tags}
\title{Prepare a list of data frames to plot repulsive tags.}
\usage{
prep_repulsive_tags(df_list)
}
\arguments{
\item{df_list}{A list of data frames.}
}
\value{
A list of dataframes, prepared for \code{\link[=plot_repulsive_tags]{plot_repulsive_tags()}}.
}
\description{
This function may be rarely necessary. It wraps other functions that you will
likely want to use directly, so you have more flexibility. The input of this
function may also be a rare data structure. It is a list of dataframes, each
corresponding to a 20x20 quadrat and with the variables: "tag", "lx", "ly",
"subtag", "status", "symbol".
}
\examples{
list_of_dataframes <- toy_list
str(list_of_dataframes)

prepared <- prep_repulsive_tags(list_of_dataframes)
str(prepared[1:2])
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_latest_tree_status}},
  \code{\link{add_quadrat_and_subquadrat_from_list}},
  \code{\link{add_quadrat_to_df_list}},
  \code{\link{add_subquad_limits}},
  \code{\link{add_subquadrat}},
  \code{\link{identify_subquadrat}}
}
