% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_for_plot_repel}
\alias{prepare_for_plot_repel}
\title{Wrap a number of functions that prepare a list of data frames for plotting.}
\usage{
prepare_for_plot_repel(df_list)
}
\description{
Wrap a number of functions that prepare a list of data frames for plotting.
}
\keyword{internal}
