% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_first}
\alias{print_first}
\title{Print one element of a list and return the list invisibly.}
\usage{
print_first(.x, element = "element")
}
\arguments{
\item{.x}{A list.}

\item{element}{A string describing each element of the input list.}
}
\value{
Returns the input invisibly.
}
\description{
Printing to screen all elements of a list may be time consuming. This
function saves time by printing only the first element of the list.
}
\details{
The element printed is extracted with \code{[} -- not with \code{[[} as in
[dplyr::first()] -- to preserve the name of the element (if any).

[\code{-- not with}[[` as in
[dplyr::first()]: R:%60%20--%20not%20with%20%60[[%60%20as%20in%0A[dplyr::first()
}
\examples{
print_first(list(1, 2))
print_first(list(1, 2), "plot")
}
\seealso{
\code{\link[dplyr:first]{dplyr::first()}}
}
