% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_dead_twice}
\alias{rm_dead_twice}
\title{Exclude trees that were found dead in the last and previous last censuses.}
\usage{
rm_dead_twice(x, cns_var = "censusid")
}
\arguments{
\item{x}{Dataframe; ViewFullTable with at least two censuses.}

\item{cns_var}{String; name of variable giving census}
}
\description{
Excludes trees that were found dead in the last and previous last censuses.
Said another way, this function keeps trees that were found in one of the
last two censuses.
}
