% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{set_axis_limits}
\alias{set_axis_limits}
\title{Set the map limits.}
\usage{
set_axis_limits(p, xlim = NULL, ylim = NULL)
}
\arguments{
\item{p}{A ggplot.}

\item{xlim, ylim}{A vector of length 2, for example \code{c(0, 500)}, giving the
minimum and maximum limits of the vertical and horizontal coordinates.}
}
\description{
Set the map limits.
}
\examples{
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_stem5_random, sp \%in\% some_sp)
p <- plot_base_census(census)
p

set_axis_limits(p, xlim = c(0, 100), ylim = c(0, 400))

fgeo.data::luquillo_elevation \%>\% 
  plot_base_elevation() \%>\% 
  set_axis_limits(xlim = c(0, 100), ylim = c(0, 400))
}
\seealso{
\code{\link[=plot_species_or_elevation]{plot_species_or_elevation()}}, \code{\link[=plot_each_species]{plot_each_species()}},
\code{\link[ggplot2:coord_fixed]{ggplot2::coord_fixed()}}.

Other functions to create or modify plot layers: \code{\link{add_elevation_contrours}},
  \code{\link{add_elevation_labels}},
  \code{\link{add_species}}, \code{\link{facet}},
  \code{\link{hide}}, \code{\link{plot_base_elevation}}
}
\concept{functions to create or modify plot layers}
