% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_elev.R
\name{sp_elev}
\alias{sp_elev}
\alias{sp}
\alias{elev}
\title{Define data containing variables \code{sp} (species) and/or \code{elev} (elevation).}
\usage{
sp_elev(sp, elev)

sp(sp)

elev(elev)
}
\arguments{
\item{x}{}
}
\description{
\itemize{
\item \code{sp()} takes a dataset with variables \code{gx}, \code{gy}, and \code{sp}.
\item \code{elev()} takes a dataset with variables \code{gx}, \code{gy}, and \code{elev}.
\item \code{sp_elev()} takes two dataset:
\itemize{
\item The first one with variables \code{gx}, \code{gy}, and \code{sp}.
\item A second one with variables \code{gx}, \code{gy}, and \code{elev}.
}
}
}
\examples{

}
