% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_edgy_dead.R
\name{str_suffix_match}
\alias{str_suffix_match}
\title{Suffix a strings where a vector exactly matches one specific string.}
\usage{
str_suffix_match(string, to_match, .match, suffix)
}
\description{
Suffix a strings where a vector exactly matches one specific string.
}
\keyword{".d"}
\keyword{".d"))}
\keyword{"01",}
\keyword{"02",}
\keyword{"alive"}
\keyword{"dead",}
\keyword{"not-dead"),}
\keyword{"tag2"),}
\keyword{#}
\keyword{)}
\keyword{<-}
\keyword{=}
\keyword{Examples}
\keyword{Status,}
\keyword{c("dead",}
\keyword{c("tag1",}
\keyword{internal}
\keyword{library(dplyr)}
\keyword{mutate(vft,}
\keyword{str_suffix_match(}
\keyword{str_suffix_match(Tag,}
\keyword{string}
\keyword{tagged}
\keyword{tribble(}
\keyword{vft}
\keyword{~Status,}
\keyword{~Tag,}
