% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_edge_tag.R
\name{suffix_edge_tag}
\alias{suffix_edge_tag}
\title{Suffix tags of dead stems that map beyond the quadrat's edge.}
\usage{
suffix_edge_tag(x, .match, suffix, x_q = 20, y_q = x_q)
}
\arguments{
\item{x}{A ViewFullTable.}

\item{.match}{A character string indicaing the matching value of the variable
status to suffix.}

\item{suffix}{A character string to suffix tree tags with.}

\item{x_q, y_q}{Integer; the x and y size in meters of a quadrat.}
}
\value{
A modified version of \code{x}.
}
\description{
Suffix tags of dead stems that map beyond the quadrat's edge.
}
\examples{
vft <- tibble::tibble(
  QX = 21:22,
  QY = 21:22,
  Tag = c("01", "02"),
  Status = c("dead", "alive")
)
vft

suffix_edge_tag(
  x = vft,
  .match = "dead",
  suffix = "_d",
  x_q = 20
)
suffix_edge_tag(
  x = vft,
  .match = "alive",
  suffix = "_a",
  x_q = 20
)
}
