% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_edgy_dead.R
\name{suffix_edgy_dead}
\alias{suffix_edgy_dead}
\title{Suffix tags of dead stems that map beyond the quadrat's edge.}
\usage{
suffix_edgy_dead(x, status_d, suffix, x_q = 20, y_q = x_q)
}
\arguments{
\item{x}{A ForestGEO table, either a ViewFullTable or a census table.}

\item{status_d}{A character string indicaing the value of the variable 
status that corresponds to dead stems.}

\item{suffix}{A character string to suffix tree tags with.}

\item{x_q, y_q}{Integer; the x and y size in meters of a quadrat.}
}
\value{
A modified version of \code{x}.
}
\description{
Suffix tags of dead stems that map beyond the quadrat's edge.
}
\examples{
stem <- bciex::bci12s7mini
# Add x and y coordinates relative to each quadrat
with_lxly <- fgeo.tool::add_lxly(stem)
# Focus on dead stems
dead <- with_lxly[with_lxly$status == "D", ]

# Suffix only dead trees that spillover beyond `10`
spillover <- 10
relevant_vars <- c("tag", "lx", "ly")
suffixed_over10 <- suffix_edgy_dead(dead, "D", "*", x_q = spillover)
suffixed_over10[1:10, relevant_vars]

# To tag all dead trees, set spillover to 0
spillover <- 0
suffixed_all <- suffix_edgy_dead(dead, "D", "*", x_q = spillover)
suffixed_all[1:10, relevant_vars]
}
