% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_quad.R
\name{tag_dead}
\alias{tag_dead}
\title{Add the ending ".d" to dead stems.}
\usage{
tag_dead(x, status)
}
\arguments{
\item{x}{A character vector, giving the tag of a stem.}

\item{status}{A character vector giving the status of a stem.}
}
\value{
A modified version of \code{x}.
}
\description{
Add the ending ".d" to dead stems.
}
\examples{
library(dplyr)
vft <- tribble(
  ~Tag, ~Status,
  "01", "dead",
  "02", "alive"
)
mutate(vft, tagged = tag_dead(Tag, Status))
}
