% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_dbh_bubbles}
\alias{theme_dbh_bubbles}
\title{Theme for \code{theme_dbh_bubbles()}.}
\usage{
theme_dbh_bubbles(axis.text = element_blank(),
  panel.background = element_rect(fill = "white"),
  panel.grid.minor = element_blank(),
  panel.grid.major = element_blank(), legend.position = "none",
  plot.title = element_text(size = 20),
  plot.subtitle = element_text(size = 12),
  panel.border = element_rect(colour = "black", fill = NA),
  axis.ticks.length = unit(-0.1, "cm"), ...)
}
\arguments{
\item{axis.text}{tick labels along axes (\code{element_text}; inherits from
\code{text})}

\item{panel.background}{background of plotting area, drawn underneath plot
(\code{element_rect}; inherits from \code{rect})}

\item{panel.grid.minor}{minor grid lines (\code{element_line}; inherits from
\code{panel.grid})}

\item{panel.grid.major}{major grid lines (\code{element_line}; inherits from
\code{panel.grid})}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{plot.title}{plot title (text appearance) (\code{element_text}; inherits
from \code{title}) left-aligned by default}

\item{plot.subtitle}{plot subtitle (text appearance) (\code{element_text};
inherits from \code{title}) left-aligned by default}

\item{panel.border}{border around plotting area, drawn on top of plot so that
it covers tick marks and grid lines. This should be used with
\code{fill=NA}
(\code{element_rect}; inherits from \code{rect})}

\item{axis.ticks.length}{length of tick marks (\code{unit})}

\item{...}{additional element specifications not part of base ggplot2. If
supplied \code{validate} needs to be set to \code{FALSE}.}
}
\description{
Theme for \code{theme_dbh_bubbles()}.
}
\examples{
class(theme_dbh_bubbles())
census <- fgeo.data::luquillo_tree5_random
plot_base_census(census) + theme_dbh_bubbles()
}
\seealso{
Other functions to plot dbh bubbles: \code{\link{header_dbh_bubbles}},
  \code{\link{plot_dbh_bubbles_by_quadrat}}
}
\concept{functions to plot dbh bubbles}
