% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_default}
\alias{theme_default}
\title{Default plot theme.}
\usage{
theme_default(p, panel.grid.minor = element_line(linetype = "dashed"),
  ...)
}
\arguments{
\item{p}{A ggplot.}

\item{panel.grid.minor}{minor grid lines (\code{element_line}; inherits from
\code{panel.grid})}

\item{...}{additional element specifications not part of base ggplot2. If
supplied \code{validate} needs to be set to \code{FALSE}.}
}
\description{
Default plot theme.
}
\examples{
census <- fgeo.data::luquillo_tree5_random
theme_default(plot_base_census(census))
}
\seealso{
Other functions to create or modify plot layers: \code{\link{add_elevation_contrours}},
  \code{\link{add_elevation_labels}},
  \code{\link{add_species}}, \code{\link{axis_limits}},
  \code{\link{facet}}, \code{\link{hide}},
  \code{\link{plot_base_elevation}}
}
\concept{functions to create or modify plot layers}
