% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_quad.R
\name{theme_map_quad}
\alias{theme_map_quad}
\title{Theme for map_quad.}
\usage{
theme_map_quad(axis.text = element_blank(),
  panel.background = element_rect(fill = "white"),
  panel.grid.minor = element_blank(), panel.grid.major = element_blank(),
  legend.position = "none", plot.title = element_text(size = 20),
  plot.subtitle = element_text(size = 12),
  panel.border = element_rect(colour = "black", fill = NA),
  axis.ticks.length = unit(-0.1, "cm"), ...)
}
\arguments{
\item{...}{Arguments passed on to \code{ggplot2::theme}
\describe{
  \item{line}{all line elements (\code{element_line})}
  \item{rect}{all rectangular elements (\code{element_rect})}
  \item{text}{all text elements (\code{element_text})}
  \item{title}{all title elements: plot, axes, legends (\code{element_text};
inherits from \code{text})}
  \item{aspect.ratio}{aspect ratio of the panel}
  \item{axis.title}{label of axes (\code{element_text}; inherits from
\code{text})}
  \item{axis.title.x}{x axis label (\code{element_text}; inherits from
\code{axis.title})}
  \item{axis.title.x.top}{x axis label on top axis (\code{element_text};
inherits from \code{axis.title.x})}
  \item{axis.title.y}{y axis label (\code{element_text}; inherits from
\code{axis.title})}
  \item{axis.title.y.right}{y axis label on right axis (\code{element_text};
inherits from \code{axis.title.y})}
  \item{axis.text}{tick labels along axes (\code{element_text}; inherits from
\code{text})}
  \item{axis.text.x}{x axis tick labels (\code{element_text}; inherits from
\code{axis.text})}
  \item{axis.text.x.top}{x axis tick labels on top axis (\code{element_text};
inherits from \code{axis.text.x})}
  \item{axis.text.y}{y axis tick labels (\code{element_text}; inherits from
\code{axis.text})}
  \item{axis.text.y.right}{y axis tick labels on right axis
(\code{element_text}; inherits from \code{axis.text.y})}
  \item{axis.ticks}{tick marks along axes (\code{element_line}; inherits from
\code{line})}
  \item{axis.ticks.x}{x axis tick marks (\code{element_line}; inherits from
\code{axis.ticks})}
  \item{axis.ticks.y}{y axis tick marks (\code{element_line}; inherits from
\code{axis.ticks})}
  \item{axis.ticks.length}{length of tick marks (\code{unit})}
  \item{axis.line}{lines along axes (\code{element_line}; inherits from
\code{line})}
  \item{axis.line.x}{line along x axis (\code{element_line}; inherits from
\code{axis.line})}
  \item{axis.line.y}{line along y axis (\code{element_line}; inherits from
\code{axis.line})}
  \item{legend.background}{background of legend (\code{element_rect}; inherits
from \code{rect})}
  \item{legend.margin}{the margin around each legend (\code{margin})}
  \item{legend.spacing}{the spacing between legends (\code{unit})}
  \item{legend.spacing.x}{the horizontal spacing between legends (\code{unit});
inherits from \code{legend.spacing}}
  \item{legend.spacing.y}{the horizontal spacing between legends (\code{unit});
inherits from \code{legend.spacing}}
  \item{legend.key}{background underneath legend keys (\code{element_rect};
inherits from \code{rect})}
  \item{legend.key.size}{size of legend keys (\code{unit})}
  \item{legend.key.height}{key background height (\code{unit}; inherits from
\code{legend.key.size})}
  \item{legend.key.width}{key background width (\code{unit}; inherits from
\code{legend.key.size})}
  \item{legend.text}{legend item labels (\code{element_text}; inherits from
\code{text})}
  \item{legend.text.align}{alignment of legend labels (number from 0 (left) to
1 (right))}
  \item{legend.title}{title of legend (\code{element_text}; inherits from
\code{title})}
  \item{legend.title.align}{alignment of legend title (number from 0 (left) to
1 (right))}
  \item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}
  \item{legend.direction}{layout of items in legends ("horizontal" or
"vertical")}
  \item{legend.justification}{anchor point for positioning legend inside plot
("center" or two-element numeric vector) or the justification according to
the plot area when positioned outside the plot}
  \item{legend.box}{arrangement of multiple legends ("horizontal" or
"vertical")}
  \item{legend.box.just}{justification of each legend within the overall
bounding box, when there are multiple legends ("top", "bottom", "left", or
"right")}
  \item{legend.box.margin}{margins around the full legend area, as specified
using \code{\link{margin}}}
  \item{legend.box.background}{background of legend area (\code{element_rect};
inherits from \code{rect})}
  \item{legend.box.spacing}{The spacing between the plotting area and the
legend box (\code{unit})}
  \item{panel.background}{background of plotting area, drawn underneath plot
(\code{element_rect}; inherits from \code{rect})}
  \item{panel.border}{border around plotting area, drawn on top of plot so that
  it covers tick marks and grid lines. This should be used with
  \code{fill=NA}
(\code{element_rect}; inherits from \code{rect})}
  \item{panel.spacing}{spacing between facet panels (\code{unit})}
  \item{panel.spacing.x}{horizontal spacing between facet panels (\code{unit};
inherits from \code{panel.spacing})}
  \item{panel.spacing.y}{vertical spacing between facet panels (\code{unit};
inherits from \code{panel.spacing})}
  \item{panel.grid}{grid lines (\code{element_line}; inherits from \code{line})}
  \item{panel.grid.major}{major grid lines (\code{element_line}; inherits from
\code{panel.grid})}
  \item{panel.grid.minor}{minor grid lines (\code{element_line}; inherits from
\code{panel.grid})}
  \item{panel.grid.major.x}{vertical major grid lines (\code{element_line};
inherits from \code{panel.grid.major})}
  \item{panel.grid.major.y}{horizontal major grid lines (\code{element_line};
inherits from \code{panel.grid.major})}
  \item{panel.grid.minor.x}{vertical minor grid lines (\code{element_line};
inherits from \code{panel.grid.minor})}
  \item{panel.grid.minor.y}{horizontal minor grid lines (\code{element_line};
inherits from \code{panel.grid.minor})}
  \item{panel.ontop}{option to place the panel (background, gridlines) over
the data layers.  Usually used with a transparent or blank
\code{panel.background}. (\code{logical})}
  \item{plot.background}{background of the entire plot (\code{element_rect};
inherits from \code{rect})}
  \item{plot.title}{plot title (text appearance) (\code{element_text}; inherits
from \code{title}) left-aligned by default}
  \item{plot.subtitle}{plot subtitle (text appearance) (\code{element_text};
inherits from \code{title}) left-aligned by default}
  \item{plot.caption}{caption below the plot (text appearance)
(\code{element_text}; inherits from \code{title}) right-aligned by default}
  \item{plot.margin}{margin around entire plot (\code{unit} with the sizes of
the top, right, bottom, and left margins)}
  \item{strip.background}{background of facet labels (\code{element_rect};
inherits from \code{rect})}
  \item{strip.placement}{placement of strip with respect to axes,
either "inside" or "outside". Only important when axes and strips are
on the same side of the plot.}
  \item{strip.text}{facet labels (\code{element_text}; inherits from
\code{text})}
  \item{strip.text.x}{facet labels along horizontal direction
(\code{element_text}; inherits from \code{strip.text})}
  \item{strip.text.y}{facet labels along vertical direction
(\code{element_text}; inherits from \code{strip.text})}
  \item{strip.switch.pad.grid}{space between strips and axes when strips are
switched (\code{unit})}
  \item{strip.switch.pad.wrap}{space between strips and axes when strips are
switched (\code{unit})}
  \item{complete}{set this to TRUE if this is a complete theme, such as
the one returned \code{by theme_grey()}. Complete themes behave
differently when added to a ggplot object. Also, when setting
\code{complete = TRUE} all elements will be set to inherit from blank
elements.}
  \item{validate}{\code{TRUE} to run validate_element, \code{FALSE} to bypass checks.}
}}
}
\value{
A ggplot2 theme.
}
\description{
Theme for map_quad.
}
\examples{
theme_map_quad()
}
\seealso{
\link[ggplot2:theme]{ggplot2::theme}
}
