% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myselect.R
\name{to0}
\alias{to0}
\title{Source of parameter z}
\usage{
to0(x, y, z)
}
\arguments{
\item{x}{Documented in x}

\item{y}{Documented in y}

\item{z}{Documented in to0}
}
\value{
Stuff
}
\description{
Source of parameter z
}
