% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_list}
\alias{toy_list}
\title{A toy list of dataframes for examples.}
\format{A list of two dataframes with 300 rows and 4 variables:
\itemize{
\item tag Integers giving fake names of tree tags.
\item lx,ly Doubles giving fake position of trees.
\item symbol Integers giving the fake status of the trees.
}}
\usage{
toy_list
}
\description{
This is a toy list of dataframes build based on data from Sinharaja provided
by Shameema.
}
\keyword{datasets}
