% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.R
\docType{package}
\name{try}
\alias{try}
\alias{try-package}
\alias{try-package}
\title{Try new things.}
\description{
Try makes if  easy to share code to try new things. It aims to draft a
solution to address a very specific issue.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/try}
  \item Report bugs at \url{https://github.com/forestgeo/try/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
