% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_data.R
\docType{data}
\name{trydata}
\alias{trydata}
\alias{one}
\alias{two}
\alias{three}
\title{all data sets}
\format{An object of class \code{numeric} of length 1.}
\usage{
one
}
\description{
all data sets
}
\keyword{datasets}
