% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{wrap}
\alias{wrap}
\title{Facet data by a variable, and wrap multiple plots to fit a single page.}
\usage{
wrap(p, facets, ...)
}
\arguments{
\item{p}{A ggplot.}

\item{facets}{Either a formula or character vector. Use either a
one sided formula, \code{~a + b}, or a character vector, \code{c("a", "b")}.}

\item{...}{Arguments passed on to \code{ggplot2::facet_wrap}
\describe{
  \item{facets}{Either a formula or character vector. Use either a
one sided formula, \code{~a + b}, or a character vector, \code{c("a", "b")}.}
  \item{nrow}{Number of rows and columns.}
  \item{ncol}{Number of rows and columns.}
  \item{scales}{should Scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"}).}
  \item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}
  \item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with formulae of the type \code{~cyl + am}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link{labeller}()}. See
\code{\link{label_value}} for more details and pointers to other
options.}
  \item{as.table}{If \code{TRUE}, the default, the facets are laid out like
a table with highest values at the bottom-right. If \code{FALSE}, the
facets are laid out like a plot with the highest value at the top-right.}
  \item{switch}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}
  \item{drop}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}
  \item{dir}{Direction: either "h" for horizontal, the default, or "v", for
vertical.}
  \item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
"bottom", "left", "right")}}
}}
}
\description{
This function behaves like \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} but it is designed to
allow using it as part of a pipe.
}
\examples{
census <- fgeo.data::luquillo_stem5_random
species <- c("PREMON", "CASARB")
two_sp <- dplyr::filter(census, sp \%in\% species)

add_sp(map_gx_gy(two_sp))

add_sp(map_gx_gy(two_sp)) \%>\% 
  wrap("sp")
# Same
add_sp(map_gx_gy(two_sp)) \%>\% 
  wrap(~sp)

add_sp(map_gx_gy(two_sp)) \%>\% 
  wrap("status")

add_sp(map_gx_gy(two_sp)) \%>\% 
  wrap(c("sp", "status"))
# Same
add_sp(map_gx_gy(two_sp)) \%>\% 
  wrap(~ sp + status)
}
\seealso{
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{hide}},
  \code{\link{label_elev}}, \code{\link{limit_gx_gy}},
  \code{\link{map_gx_gy_elev}}
}
\concept{map components.}
