% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_latest_tree_status}
\alias{add_latest_tree_status}
\title{Add alternatives to symbol, that are easier to understand.}
\usage{
add_latest_tree_status(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
The modified data frame.
}
\description{
Add alternatives to symbol, that are easier to understand.
}
\examples{
\dontrun{
library(purrr)
library(try)
sin_q20 <- purrr::map(sin_q20, add_latest_tree_status)
}
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_limits_shrinked}},
  \code{\link{add_limits}},
  \code{\link{add_quadrat_and_subquadrat_from_list}},
  \code{\link{add_quadrat_to_df_list}},
  \code{\link{add_subquadrat}},
  \code{\link{identify_subquadrat}},
  \code{\link{prep_repulsive_tags}}
}
\keyword{internal}
