% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_latest_tree_status}
\alias{add_latest_tree_status}
\title{Add alternatives to the variable \code{symbol} that are easier to understand.}
\usage{
add_latest_tree_status(df)
}
\arguments{
\item{df}{A data frame with a variable named \code{symbol}.}
}
\value{
The modified data frame.
}
\description{
The variable \code{symbol} codes the status of a tree in a very succint way. This
function unpacks the meaning of \code{symbol}. Be careful: this function is
specifically designed for one particular dataset (from Sinharaja) and if you
use it on other data sets it will likely produce wrong results (and the
function may not throw an error).
}
\examples{
 # Add to a single dataframe
 df <- toy_list[[1]]
 head(add_latest_tree_status(df))

 # Add to each dataframe in a list
 df_list <- toy_list
 result <- lapply(df_list, add_latest_tree_status)
 # Just show a few rows of each dataframe
 lapply(result, head)
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_quadrat_and_subquadrat_from_list}},
  \code{\link{add_quadrat_to_df_list}},
  \code{\link{add_subquad_limits}},
  \code{\link{prep_repulsive_tags}}
}
