% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_limits}
\alias{add_limits}
\title{Add plot limits for quadrats 1-4.}
\usage{
add_limits(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A modified data frame.
}
\description{
Add plot limits for quadrats 1-4.
}
\keyword{internal}
