% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_quadrat_and_subquadrat_from_list}
\alias{add_quadrat_and_subquadrat_from_list}
\title{Help identify_subquadrat().}
\usage{
add_quadrat_and_subquadrat_from_list(df_list, ...)
}
\arguments{
\item{df_list}{A list of dataframes}

\item{...}{Arguments passed to \code{\link[=add_subquadrat]{add_subquadrat()}}. Importantly, this allows
changing the size of the subquadrats.}
}
\description{
Help identify_subquadrat().
}
\examples{
df_list <- toy_list[1:2]
# Extracting first quadrat and showing the head of each subquadrat
added <- add_quadrat_and_subquadrat_from_list(df_list)[[1]]
lapply(added, head)
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_latest_tree_status}},
  \code{\link{add_quadrat_to_df_list}},
  \code{\link{add_subquad_limits}},
  \code{\link{add_subquadrat}},
  \code{\link{identify_subquadrat}},
  \code{\link{prep_repulsive_tags}}
}
