% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{add_quadrat_to_df_list}
\alias{add_quadrat_to_df_list}
\title{Add the name of each element of a list as a value in the variable quadrat}
\usage{
add_quadrat_to_df_list(df)
}
\arguments{
\item{df}{A list of data frames}
}
\value{
A list
}
\description{
Add the name of each element of a list as a value in the variable quadrat
}
\examples{
# xxx add example
}
