% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_quadrat_to_df_list}
\alias{add_quadrat_to_df_list}
\title{Add the name of each element of a list as a value in the variable quadrat}
\usage{
add_quadrat_to_df_list(df_list)
}
\arguments{
\item{df_list}{A list of data frames.}
}
\value{
A list of dataframes with the variable quadrat.
}
\description{
Add the name of each element of a list as a value in the variable quadrat
}
\examples{
\dontrun{
df_list <- toy_list[1:2]
with_quadrat <- add_quadrat_to_df_list(df_list)
str(with_quadrat)
}
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_latest_tree_status}},
  \code{\link{add_quadrat_and_subquadrat_from_list}},
  \code{\link{add_subquad_limits}}
}
