% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{add_sp}
\alias{add_sp}
\title{Represent species with points.}
\usage{
add_sp(p, data = NULL, fill = "sp", shape = 21, point_size = 3)
}
\arguments{
\item{p}{A ggplot.}

\item{data}{A dataframe; passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{fill}{Character; either a colour or "sp", which maps each species to a
different color.}

\item{shape, point_size}{A number giving point size and shape (as in
\code{\link[graphics:points]{graphics::points()}}). Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
}
\description{
Represent species with points.
}
\examples{
# Small dataset with a few species for quick examples
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% some_sp)
p <- map_gx_gy(census)

add_sp(p)

add_sp(p, fill = "white", shape = 22, point_size = 4)
}
\seealso{
\code{\link[=map_sp_elev]{map_sp_elev()}}, \code{\link[=plot_each_species]{plot_each_species()}}, \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.

Other map components.: \code{\link{contour_elev}},
  \code{\link{hide}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
\concept{map components.}
