% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{add_sp}
\alias{add_sp}
\title{Represent species with points.}
\usage{
add_sp(p, data = NULL, fill = "sp", shape = 21, point_size = 3)
}
\arguments{
\item{p}{A ggplot.}

\item{data}{A dataframe; passed to [ggplot2::ggplot()].}

\item{fill}{Character; either a colour or "sp", which maps each species to a
different color.}

\item{shape, point_size}{A number giving point size and shape (as in
[graphics::points()]). Passed to [ggplot2::geom_point()].}
}
\description{
Represent species with points.
}
\seealso{
Other map components.: \code{\link{contour_elev}},
  \code{\link{hide}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
