% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_status_tree}
\alias{add_status_tree}
\title{Ensure that the status refers to the tree, not to the stem.}
\usage{
add_status_tree(df)
}
\arguments{
\item{df}{A ViewFullTable or a fgeo table.}
}
\value{
The input data set with lowercase names, and with the additional
variable status_tree.
}
\description{
Ensure that the status refers to the tree, not to the stem.
}
