% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subquadrat.R
\name{add_subquadrat}
\alias{add_subquadrat}
\title{Add a quadrat variable to a dataframe based based on QX and QY coordinates.}
\usage{
add_subquadrat(df, dim_x, dim_y, divide_x, divide_y)
}
\arguments{
\item{df}{A dataframe.}

\item{dim_x, dim_y}{Quadrat dimension for the plot. For example, in Sinharaja
both are 20.}

\item{divide_x, divide_y}{Total number of divisions of each quadrat side. For
example, For Sinharaja and most other plots, the value of these arguments
is 5, which results in a 4x4 grid of subquadrats within each quadrat.}
}
\value{
A dataframe with the additional variable \code{subquadrat}.
}
\description{
Add a quadrat variable to a dataframe based based on QX and QY coordinates.
}
\examples{
df <- sinharaja::sinh_vftbl_selected
with_subquadrat <- add_subquadrat(df, 20, 20, 5, 5)
head(with_subquadrat)
}
