% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{add_subquadrat}
\alias{add_subquadrat}
\title{Adds subquadrat variable to a data frame, using Shameema's code.}
\usage{
add_subquadrat(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A list of 4 data frames.
}
\description{
Adds subquadrat variable to a data frame, using Shameema's code.
}
\examples{
df <- sin_q20[[15]]
lapply(add_subquadrat(df), head)
}
\author{
Shameema Jafferjee Esufali \href{mailto:shameemaesufali@gmail.com}{shameemaesufali@gmail.com}.
}
