% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tag.R
\name{add_subquadrat}
\alias{add_subquadrat}
\title{Add a quadrat variable to a dataframe based based on qx and qy coordinates.}
\usage{
add_subquadrat(df, x_q, y_q = x_q, x_sq, y_sq = x_sq)
}
\arguments{
\item{df}{A ViewFullTable dataframe.}

\item{x_q, y_q}{Size in meters of a quadrat's side. For ForestGEO sites, a
common value is 20.}

\item{x_sq, y_sq}{Size in meters of a subquadrat's side. For ForestGEO-CTFS sites, a
common value is 5.}
}
\value{
A dataframe with the additional variable \code{subquadrat}.
}
\description{
Add a quadrat variable to a dataframe based based on qx and qy coordinates.
}
\examples{
\dontrun{
# ngelnyaki is a private dataset
df <- ngelnyaki::ngelnyaki_vft_unid
with_subquadrat <- add_subquadrat(df, 20, 20, 5, 5)
head(with_subquadrat[c("qx", "qy", "subquadrat")])
}
}
\author{
Anudeep Singh.
}
