% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-sp_elev.R
\name{autoplot.elev}
\alias{autoplot.elev}
\title{Quick plot of species distribution.}
\usage{
\method{autoplot}{elev}(object, ...)
}
\arguments{
\item{object}{An object created with \code{elev()}.}

\item{...}{Other arguments passed to \code{add_elevation_contours()}.}
}
\value{
A "ggplot".
}
\description{
Automatically plot the \code{elev} variable of a ForestGEO-like dataset of class
'elev'. You can create an 'elev' \code{object} with:

\preformatted{object <- elev(DATA-WITH-VARIABLE-elev)
}

See sections \strong{Usage} and \strong{Examples}.
}
\examples{
elevation_list <- fgeo.data::luquillo_elevation
autoplot(elev(elevation_list))
# Same
elevation_dataframe <- elevation_list$col
autoplot(elev(elevation_dataframe))

# Customize
p <- autoplot(elev(elevation_dataframe), fill = "red")
p
hide_color_legend(p)
}
\seealso{
\code{\link[=autoplot]{autoplot()}}, \code{\link[=elev]{elev()}}, \link{add_elevation_contours}.

Other autoplots: \code{\link{autoplot.fgeo_habitat}},
  \code{\link{autoplot.sp_elev}}, \code{\link{autoplot.sp}}
}
\concept{autoplots}
