% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.fgeo_habitat.R
\name{autoplot.fgeo_habitat}
\alias{autoplot.fgeo_habitat}
\title{Plot objects of class "fgeo_habitat".}
\usage{
\method{autoplot}{fgeo_habitat}(object, ...)
}
\arguments{
\item{object}{An S3 object of class "fgeo_habitat".}

\item{...}{Other arguments passed to methods.}
}
\value{
An object of class "ggplot".
}
\description{
Plot objects of class "fgeo_habitat".
}
\examples{
elev_list <- fgeo.data::luquillo_elevation
habitats <- fgeo.tool::fgeo_habitat(elev_list, gridsize = 20, n = 4)

autoplot(habitats)
}
\seealso{
Other autoplots: \code{\link{autoplot.sp}}
}
\concept{autoplots}
