% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.fgeo_habitat.R
\name{autoplot.fgeo_habitat}
\alias{autoplot.fgeo_habitat}
\title{Plot habitats.}
\usage{
\method{autoplot}{fgeo_habitat}(object, ...)
}
\arguments{
\item{object}{An object of class "fgeo_habitat" (see \code{fgeo_habitat} at
\url{https://forestgeo.github.io/fgeo/articles/siteonly/reference.html}).}

\item{...}{Not used (included for compatibility across methods).}
}
\value{
An object of class "ggplot".
}
\description{
Plot habitats.
}
\examples{
assert_is_installed("fgeo.x")
habitats <- fgeo.x::habitat
autoplot(habitats)
}
\seealso{
Other plot functions: 
\code{\link{autoplot.sp_elev}()},
\code{\link{autoplot_by_species.sp_elev}()},
\code{\link{elev}()},
\code{\link{plot_dbh_bubbles_by_quadrat}()},
\code{\link{plot_tag_status_by_subquadrat}()},
\code{\link{sp_elev}()},
\code{\link{sp}()}

Other autoplots: 
\code{\link{autoplot.sp_elev}()},
\code{\link{elev}()},
\code{\link{sp_elev}()},
\code{\link{sp}()}
}
\concept{autoplots}
\concept{plot functions}
