% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-sp_elev.R
\name{autoplot.sp}
\alias{autoplot.sp}
\title{Quick plot of species distribution.}
\usage{
\method{autoplot}{sp}(object, ...)
}
\arguments{
\item{object}{An object created with \code{sp()}.}

\item{...}{Other arguments passed to \code{add_species()}.}
}
\value{
A "ggplot".
}
\description{
Automatically plot the \code{sp} variable of a ForestGEO-like dataset of class
'sp'. You can create a 'sp' \code{object} with:

\preformatted{object <- sp(DATA-WITH-VARIABLE-sp)
}

See sections \strong{Usage} and \strong{Examples}.
}
\examples{
# Small dataset with a few species for quick examples
tree <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% c("PREMON", "CASARB"))
autoplot(sp(tree))

# Customize
autoplot(sp(tree), point_size = 1)
}
\seealso{
\code{\link[=autoplot]{autoplot()}}, \code{\link[=sp]{sp()}}, \code{\link[=add_species]{add_species()}}.

Other autoplots: \code{\link{autoplot.elev}},
  \code{\link{autoplot.fgeo_habitat}},
  \code{\link{autoplot.sp_elev}}
}
\concept{autoplots}
