% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-sp_elev.R
\name{autoplot.sp}
\alias{autoplot.sp}
\title{Quick plot of species distribution.}
\usage{
\method{autoplot}{sp}(object, fill = "black", shape = 21,
  point_size = 3, facet = TRUE, hide_color_legend = FALSE,
  xlim = NULL, ylim = NULL, custom_theme = NULL)
}
\arguments{
\item{object}{An object created with \code{sp()}.}

\item{fill}{Character; either a colour or "sp", which maps each species to a
different color.}

\item{shape, point_size}{A number giving point size and shape (as in
\code{\link[graphics:points]{graphics::points()}}). Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{facet}{(Not available for \code{plot_each_species()}) Logical; \code{TRUE} wraps
multiple maps within the area of a single graphic plot.}

\item{hide_color_legend}{Logical; \code{TRUE} hides the color legend.}

\item{xlim}{A vector of length 2, for example \code{c(0, 500)}, giving the
minimum and maximum limits of the vertical and horizontal coordinates.}

\item{ylim}{A vector of length 2, for example \code{c(0, 500)}, giving the
minimum and maximum limits of the vertical and horizontal coordinates.}

\item{custom_theme}{A valid \code{\link[ggplot2:theme]{ggplot2::theme()}}. \code{NULL} uses the default
theme \code{\link[=theme_default]{theme_default()}}.}
}
\value{
A "ggplot".
}
\description{
Automatically plot the \code{sp} variable of a ForestGEO-like dataset of class
'sp'. You can create a 'sp' \code{object} with:

\preformatted{object <- sp(DATA-WITH-VARIABLE-sp)
}

See sections \strong{Usage} and \strong{Examples}.
}
\section{Comparison with ggplot2}{

These functions wrap functions from the \strong{ggplot2} package. For more
control you can use \strong{ggplot2} directly or smaller wrappers in
\strong{fgeo.map} (see the sections See Also and Examples). \strong{ggplot2}
will give you maximum control and is your best choice if you are already
familiar with it. The wrapper functions in \strong{fgeo.map} focus on
ForestGEO's data: For example, their names commonly mention the variable of a
ForestGEO dataset they work with.

The most important difference between \strong{ggplot2} and \strong{fgeo.map}
is in the way you compose multiple functions to create complex plots. With
\strong{ggplot2} you compose multiple functions with \code{+}. For example,
to apply functions \code{f()} then \code{g()} you do something like this:
\code{f(data) + g()}. With \strong{fgeo.map} you compose multiple functions
in the usual way with \code{g(f(data))}, or in the modern way with the pipe:
\code{f(data)} \code{\%>\%} \code{g()} -- it is up to you. To learn more
about the pipe see http://r4ds.had.co.nz/pipes.html.
}

\examples{
# Small dataset with a few species for quick examples
tree <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% c("PREMON", "CASARB"))
autoplot(sp(tree))

# Customize
autoplot(sp(tree), point_size = 1)
}
\seealso{
\code{\link[=autoplot]{autoplot()}}, \code{\link[=sp]{sp()}}, \code{\link[=add_species]{add_species()}}.

Other autoplots: \code{\link{autoplot.elev}},
  \code{\link{autoplot.fgeo_habitat}},
  \code{\link{autoplot.sp_elev}},
  \code{\link{plot_by_species}}
}
\concept{autoplots}
