% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapply_sp_elev.R
\name{contour_elev}
\alias{contour_elev}
\title{Represent elevation with lines.}
\usage{
contour_elev(p, contour_size = 1, low = "blue", high = "red",
  bins = bins)
}
\description{
Represent elevation with lines.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{fgeo_facets}},
  \code{\link{hide_axis_labels}},
  \code{\link{hide_legend_elev}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}}
}
