% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{contour_elev}
\alias{contour_elev}
\title{Represent elevation with lines.}
\usage{
contour_elev(p, contour_size = 1, low = "blue", high = "red",
  bins = NULL)
}
\arguments{
\item{p}{A ggplot.}

\item{contour_size}{A number giving the size of the contour of elevation
lines. Passed to \code{ggplot2::stat_contour()} (see \code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}).}

\item{low, high}{A string giving a color of the elevation lines representing
low and high elevation.}

\item{bins}{A number giving the number of elevation lines to map.}
}
\description{
Represent elevation with lines.
}
\examples{
p <- map_gx_gy_elev(fgeo.data::luquillo_elevation)
contour_elev(p)
contour_elev(p, contour_size = 0.5, low = "grey", high = "black", bins = 4)
}
\seealso{
\code{\link[=plot_each_species]{plot_each_species()}}, \code{\link[=map_sp_elev]{map_sp_elev()}}, \code{\link[=map_elev]{map_elev()}},
\code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{hide}}, \code{\link{label_elev}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
\concept{map components.}
