% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{top4quad}
\alias{top1quad}
\title{Data sets for examples; adapted from \link[bciex:bci12vft_mini]{bciex::bci12vft_mini}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 964 rows and 28 columns.}
\usage{
top4quad

top1quad
}
\description{
Data sets for examples; adapted from \link[bciex:bci12vft_mini]{bciex::bci12vft_mini}.
}
\examples{
library(dplyr)
all_quad <- bciex::bci12vft_mini \%>\%
  dplyr::rename(QX = x, QY = y) \%>\%
  dplyr::filter(PlotID == 1) \%>\%
  fgeo.tool::row_top(CensusID)

top4quad <- all_quad \%>\% fgeo.tool::row_top(QuadratID, 4)
identical(top4quad, fgeo.map::top4quad)

top1quad <- all_quad \%>\% fgeo.tool::row_top(QuadratID, 1)
identical(top1quad, fgeo.map::top4quad)

top1quad
}
\keyword{datasets}
