% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exists_in_pkg}
\alias{exists_in_pkg}
\title{Test if an object exists in the namespace of a package.}
\usage{
exists_in_pkg(object, package)
}
\arguments{
\item{object}{String giving the object to find.}

\item{package}{String giving the package which namespace to search.}
}
\value{
Logical.
}
\description{
Test if an object exists in the namespace of a package. Useful inside an
\code{if} statement.
}
\seealso{
Other Functions for developers.: \code{\link{check_crucial_names}},
  \code{\link{check_unique}}
}
