% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exists_in_pkg}
\alias{exists_in_pkg}
\title{Test if an object existe in the namespace of a package.}
\usage{
exists_in_pkg(object, package)
}
\arguments{
\item{object}{String giving the object to find.}

\item{package}{String giving the package which namespace to search.}
}
\value{
Logical.
}
\description{
Test if an object existe in the namespace of a package.
}
