% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot_two.R
\name{fgeo_facets}
\alias{fgeo_facets}
\alias{facet_wrap_sp}
\alias{facet_h_sp}
\alias{facet_v_sp}
\title{Facets to add on top of ggplots based on ForestGEO's data.}
\usage{
facet_wrap_sp(p, ...)

facet_h_sp(p, ...)

facet_v_sp(p, ...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.}
}
\description{
Facets to add on top of ggplots based on ForestGEO's data.
}
\seealso{
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.
}
