% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_header}
\alias{get_header}
\title{A simple way to produce a three-lines header for a plot of repulsive tags.}
\usage{
get_header(line1 = pad(c("Checking: ", "Checked date: ")),
  line2 = pad(c("Recording: "), 24), line3 = pad(c("Measuring: ",
  "Measurement date: ")))
}
\arguments{
\item{line1, line2, line3}{Strings to use on the first, second and third line
of the header.}
}
\value{
A three-lines string to be used as the header on a plot of repulsive
tags.
}
\description{
A simple way to produce a three-lines header for a plot of repulsive tags.
}
\examples{
get_header()
get_header(line1 = "Hello")
}
