% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_header}
\alias{get_header}
\title{A simple way to produce a three-lines header for \code{map_tag()}.}
\usage{
get_header(line1 = "Checking: _______________ Checked date: _______________",
  line2 = "Recording: _____________",
  line3 = "Measuring: _____________ Measurement date: _____________")
}
\arguments{
\item{line1, line2, line3}{Strings to use on the first, second and third line
of the header.}
}
\value{
A three-lines string to be used as the header on a plot of repulsive
tags.
}
\description{
A simple way to produce a three-lines header for \code{map_tag()}.
}
\examples{
# Use common defaults
get_header()

# Or make your own strings.
get_header(line1 = "This is", line2 = "a customized", line3 = "header")
get_header(line1 = "Only", line2 = "two lines", line3 = NULL)
}
\seealso{
Other functions to fine tune a plot of repulsive tags.: \code{\link{get_theme}}
}
