% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_theme}
\alias{get_theme}
\title{A theme that is customized for \code{map_tag()}.}
\usage{
get_theme(panel.grid.major = element_line(colour = "black"),
  panel.grid.minor = element_line(colour = "black", linetype = "dotted"),
  panel.background = element_rect(fill = "white"),
  plot.title = element_text(size = 20), plot.subtitle = element_text(size =
  12), legend.position = "top", legend.title = element_blank(),
  axis.ticks = element_blank(), axis.text = element_text(size = 12), ...)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{panel_grid_major_colour}{Colour of major grid
lines.}

\item{panel_grid_minor_colour, panel_grid_minor_linetype}{Colour and line type
of minor grid lines.}

\item{panel_background_fill}{Fill colour of the background of plotting area,
drawn underneath plot.}

\item{plot_title_size, plot_subtitle_size}{Size of the plot title and
subtitle.}

\item{legend_position}{The position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector).}

\item{axis_ticks}{Tick marks along axes.}

\item{axis_text_size}{Size of tick labels along axes.}

\item{legend_title}{title of legend.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} customized for \code{\link[=map_tag]{map_tag()}}.
}
\description{
Theme customized for \code{\link[=map_tag]{map_tag()}}. You can further customize the
theme by passing more arguments to \code{\link[ggplot2:theme]{ggplot2::theme()}} (via \code{...}) or by
replacing this theme completely with a new \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\examples{
ggplot2::theme()
map::get_theme()
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
