% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_theme}
\alias{get_theme}
\title{A theme that is customized for \code{map_tag()}.}
\usage{
get_theme(panel_grid_major_colour = "black",
  panel_grid_minor_colour = "black", panel_grid_minor_linetype = "dotted",
  panel_background_fill = "white", plot_title_size = 20,
  plot_subtitle_size = 12, legend_position = "top",
  axis_ticks = element_blank(), axis_text_size = 12,
  legend_title = element_blank(), ...)
}
\arguments{
\item{panel_grid_major_colour}{Colour of major grid
lines.}

\item{panel_grid_minor_colour, panel_grid_minor_linetype}{Colour and line type
of minor grid lines.}

\item{panel_background_fill}{Fill colour of the background of plotting area,
drawn underneath plot.}

\item{plot_title_size, plot_subtitle_size}{Size of the plot title and
subtitle.}

\item{legend_position}{The position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector).}

\item{axis_ticks}{Tick marks along axes.}

\item{axis_text_size}{Size of tick labels along axes.}

\item{legend_title}{title of legend.}

\item{...}{Other arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} customized for \code{\link[=map_tag]{map_tag()}}.
}
\description{
Theme customized for \code{\link[=map_tag]{map_tag()}}. You can further customize the
theme by passing more arguments to \code{\link[ggplot2:theme]{ggplot2::theme()}} (via \code{...}) or by
replacing this theme completely with a new \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\examples{
ggplot2::theme()
try::get_theme()
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.

Other functions to fine tune a plot of repulsive tags.: \code{\link{get_header}},
  \code{\link{pad}}
}
