% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak_plot.R
\name{get_theme}
\alias{get_theme}
\title{Theme passed to plot_repel().}
\usage{
get_theme()
}
\description{
Theme passed to \code{\link[=plot_repel]{plot_repel()}}. This function wraps multiple functions that
change the looks of a plot. To change the looks of a plot, change the
relevant \code{get_*()} function.
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
