% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fun_settings.R
\name{headers}
\alias{headers}
\alias{map_tag_header}
\alias{map_quad_header}
\title{Pre-made headers.}
\usage{
map_tag_header()

map_quad_header(lang = "english")
}
\arguments{
\item{lang}{= String; Language of the header: Either "english" or "spanish".}
}
\description{
Pre-made headers.
}
\examples{
cat(map_tag_header())
cat(map_quad_header())
}
