% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fun_settings.R
\name{headers}
\alias{headers}
\alias{header_tag_status}
\alias{header_dbh_bubles}
\title{Pre-made headers.}
\usage{
header_tag_status()

header_dbh_bubles(lang = "english")
}
\arguments{
\item{lang}{= String; Language of the header: Either "english" or "spanish".}
}
\description{
Pre-made headers.
}
\examples{
cat(header_tag_status())
cat(header_dbh_bubles())
}
\seealso{
Other functions to tweak plots: \code{\link{suffix_tags_beyond_edge}},
  \code{\link{themes}}
}
\concept{functions to tweak plots}
