% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{hide}
\alias{hide}
\alias{hide_axis_labels}
\alias{hide_color_legend}
\title{Hide elements of a plot.}
\usage{
hide_axis_labels(p)

hide_color_legend(p)
}
\arguments{
\item{p}{A ggplot.}
}
\description{
\code{hide_axis_labels} hides axis labels.
\code{hide_color_legend} hides the color legend, for example, of elevation lines.
}
\examples{
elevation <- fgeo.data::luquillo_elevation
p <- plot_base_elevation(elevation)
hide_axis_labels(p)

elevation \%>\% 
  plot_base_elevation() \%>\% 
  add_elevation_contrours() \%>\% 
  hide_color_legend()
}
\seealso{
\code{\link[ggplot2:labs]{ggplot2::labs()}},  \code{\link[ggplot2:guides]{ggplot2::guides()}}.

Other functions to create or modify plot layers: \code{\link{add_elevation_contrours}},
  \code{\link{add_elevation_labels}},
  \code{\link{add_species}}, \code{\link{facet}},
  \code{\link{plot_base_elevation}},
  \code{\link{set_axis_limits}}
}
\concept{functions to create or modify plot layers}
