% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{identify_subquadrat}
\alias{identify_subquadrat}
\title{Identify quadrat and subquadrat, from list to data frame.}
\usage{
identify_subquadrat(df_list)
}
\arguments{
\item{df_list}{A list of data frames.}
}
\value{
A data frames with new variables identifying quadrat, and subquadrat.
}
\description{
Identify quadrat and subquadrat, from list to data frame.
}
\examples{
\dontrun{
df_list1 <- sin_q20[1:2]
names(df_list1)
identify_subquadrat(df_list1)
}
}
\seealso{
Other functions to prepare data to plot repulsive tags.: \code{\link{add_latest_tree_status}},
  \code{\link{add_limits_shrinked}},
  \code{\link{add_limits}},
  \code{\link{add_quadrat_and_subquadrat_from_list}},
  \code{\link{add_quadrat_to_df_list}},
  \code{\link{add_subquadrat}},
  \code{\link{prep_repulsive_tags}}
}
\keyword{internal}
