% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_subquadrat.R
\name{identify_subquadrat}
\alias{identify_subquadrat}
\title{Identify quadrat and subquadrat, from list to data frame.}
\usage{
identify_subquadrat(df_list)
}
\arguments{
\item{df_list}{A list of data frames.}
}
\value{
A data frames with new variables identifying quadrat, and subquadrat.
}
\description{
Identify quadrat and subquadrat, from list to data frame.
}
\examples{
df_list1 <- sin_q20[1:2]
names(df_list1)
identify_subquadrat(df_list1)
}
