% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{label_elev}
\alias{label_elev}
\title{Label elevation lines.}
\usage{
label_elev(p, label_size = 3, label_color = "grey", xyjust = 1,
  fontface = "italic")
}
\arguments{
\item{p}{A ggplot.}

\item{label_size, label_color, fontface}{A number (`label_size`) or character
string (`label_color` and `fontface`) giving the size, colour and fontface
of the text labels for the elevation lines.}

\item{xyjust}{A number to adjust the position of the text labels of the
elevation lines.}
}
\description{
Label elevation lines.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{hide}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
