% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_each_species.R
\name{label_elev}
\alias{label_elev}
\title{Label elevation lines.}
\usage{
label_elev(p, label_size = 3, label_color = "grey", xyjust = 1,
  fontface = "italic")
}
\arguments{
\item{p}{A ggplot.}

\item{label_size, label_color, fontface}{A number (\code{label_size}) or character
string (\code{label_color} and \code{fontface}) giving the size, colour and fontface
of the text labels for the elevation lines.}

\item{xyjust}{A number to adjust the position of the text labels of the
elevation lines.}
}
\description{
Label elevation lines.
}
\examples{
elevation <- fgeo.data::luquillo_elevation

contour_elev(map_gx_gy_elev(elevation))

elevation \%>\%
  map_gx_gy_elev() \%>\% 
  contour_elev() \%>\%
  label_elev(label_size = 2, label_color = "black", xyjust = -0.25)
}
\seealso{
\code{\link[=plot_each_species]{plot_each_species()}}, \code{\link[=plot_species_or_elevation]{plot_species_or_elevation()}}, \code{\link[=plot_elevation]{plot_elevation()}},
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{hide}},
  \code{\link{limit_gx_gy}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
\concept{map components.}
