% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lapply_plot_repel}
\alias{lapply_plot_repel}
\alias{plot_repel}
\title{Plot trees in subquadrats, avoiding tree tags to overlap.}
\usage{
lapply_plot_repel(df_list, site_name = get_site_name())

plot_repel(df, site_name)
}
\arguments{
\item{df_list}{A list of data frames.}

\item{df}{A data frame as wranglec by \code{lapply_plot_repel()}.}
}
\value{
A list of plots that can be wrapped around pdf(onefile = TRUE).
}
\description{
Plot trees in subquadrats, avoiding tree tags to overlap.
}
