% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{limit_gx_gy}
\alias{limit_gx_gy}
\title{Set the map limits.}
\usage{
limit_gx_gy(p, xlim = NULL, ylim = NULL)
}
\arguments{
\item{p}{A ggplot.}

\item{xlim, ylim}{A vector of lenght 2, for example `c(0, 500)`, giving the
minimum and maximum limits of the vertical and horizontal coordinates.}
}
\description{
Set the map limits.
}
\examples{
census <- bciex::bci12s7mini
map <- map_gx_gy(census)
limit_gx_gy(map, xlim = c(0, 500), c(0, 750))

elevation <- bciex::bci_elevation
map <- map_gx_gy_elev(elevation)
limit_gx_gy(map, xlim = c(0, 500), c(0, 750))
}
\seealso{
\code{\link[=map_sp_elev]{map_sp_elev()}}, \code{\link[=maply_sp_elev]{maply_sp_elev()}}, \code{\link[ggplot2:coord_fixed]{ggplot2::coord_fixed()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{hide}},
  \code{\link{label_elev}}, \code{\link{map_gx_gy_elev}},
  \code{\link{wrap}}
}
