% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapply_sp_elev.R
\name{limit_gx_gy}
\alias{limit_gx_gy}
\title{Set the map limits.}
\usage{
limit_gx_gy(p, xlim = NULL, ylim = NULL)
}
\description{
Set the map limits.
}
\seealso{
Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{fgeo_facets}},
  \code{\link{hide_axis_labels}},
  \code{\link{hide_legend_elev}}, \code{\link{label_elev}},
  \code{\link{map_gx_gy_elev}}
}
