% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{map_gx_gy_elev}
\alias{map_gx_gy_elev}
\alias{map_gx_gy}
\title{Map a base over which other map components can later be added.}
\usage{
map_gx_gy_elev(data)

map_gx_gy(data)
}
\arguments{
\item{data}{A dataframe; passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}
}
\description{
Map a base over which other map components can later be added.
}
\examples{
# Small dataset with a few species for quick examples
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_tree5_random, sp \%in\% some_sp)
elevation <- fgeo.data::luquillo_elevation

# These functions look alike but internally they are not.
base_census <- map_gx_gy(census)
base_census
base_elevation <- map_gx_gy_elev(elevation)
base_elevation

# Use map_gx_gy() as a base for layers that need census data
add_sp(base_census)
# This fails because the base plot has no information about elevation
\dontrun{
  contour_elev(base_census)
}

# Use map_gx_gy_elev() as a base for layers that need elevation data
contour_elev(base_elevation)
# This fails because the base plot has no information about species
\dontrun{
  add_sp(base_elevation)
}
}
\seealso{
\code{\link[=map_sp_elev]{map_sp_elev()}}, \code{\link[=map_elev]{map_elev()}}, \code{\link[=plot_each_species]{plot_each_species()}},
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{hide}},
  \code{\link{label_elev}}, \code{\link{limit_gx_gy}},
  \code{\link{wrap}}
}
\concept{map components.}
