% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maply_sp_elev.R
\name{map_gx_gy_elev}
\alias{map_gx_gy_elev}
\alias{map_gx_gy}
\title{Map a base over which other map components can later be added.}
\usage{
map_gx_gy_elev(data)

map_gx_gy(data)
}
\arguments{
\item{data}{A dataframe; passed to [ggplot2::ggplot()].}
}
\description{
Map a base over which other map components can later be added.
}
\examples{
census <- fgeo.tool::top(bciex::bci12s7mini, sp, 2)
elevation <- bciex::bci_elevation

# Look alike but internally they are not. 
base_elev <- map_gx_gy_elev(elevation)
base_census <- map_gx_gy(census)

# Works
contour_elev(base_elev)

# # Fails.
# \dontrun{
# add_sp(base_elev)
# contour_elev(base_census)
# }

# Supply the correct base
add_sp(base_census)
contour_elev(base_elev)
}
\seealso{
\code{\link[=map_sp_elev]{map_sp_elev()}}, \code{\link[=map_elev]{map_elev()}}, \code{\link[=maply_sp_elev]{maply_sp_elev()}},
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.

Other map components.: \code{\link{add_sp}},
  \code{\link{contour_elev}}, \code{\link{hide}},
  \code{\link{label_elev}}, \code{\link{limit_gx_gy}},
  \code{\link{wrap}}
}
