% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_quad.R
\name{map_quad}
\alias{map_quad}
\title{Map trees within a quadrat.}
\usage{
map_quad(vft, title_quad = "Site Name, YYYY, Quadrat:",
  header = map_quad_header(), theme = theme_map_quad(), lim_min = 0,
  lim_max = 20, subquadrat_side = 5, tag_size = 2, extend_grid = 0)
}
\arguments{
\item{vft}{A dataframe -- specifically, a ForestGEO ViewFullTable.}

\item{title_quad}{A string to use as a title.}

\item{header}{A string to use as a header (see [headers]).}

\item{theme}{An object of class theme.}

\item{lim_min, lim_max}{Minimum and maximum limits of the plot area.}

\item{subquadrat_side}{Length in meters of the side of a subquadrat.}

\item{tag_size}{A number giving tag size. Passed to
[ggrepel::geom_text_repel].}

\item{extend_grid}{A number to adjust the extension of the grid lines beyond
the plot limits.}
}
\value{
A list which each element is a plot of class ggplot.
}
\description{
Map trees within a quadrat.
}
\examples{
\dontrun{
p <- map_quad(four_quadrats)
dplyr::first(p)

# Printing to .pdf optimized for size letter
pdf("default-map.pdf", paper = "letter", height = 10.5, width = 8)
p
dev.off()
}
}
