% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_quad.R
\name{map_quad}
\alias{map_quad}
\title{Map trees within a quadrat.}
\usage{
map_quad(vft, title_quad = "Site Name, YYYY, Quadrat:",
  header = map_quad_header(), theme = theme_map_quad(), lim_min = 0,
  lim_max = 20, subquadrat_side = 5, tag_size = 2, move_edge = 0)
}
\arguments{
\item{vft}{A dataframe -- specifically, a ForestGEO ViewFullTable.}

\item{title_quad}{A string to use as a title.}

\item{header}{A string to use as a header (see [headers]).}

\item{theme}{An object of class theme.}

\item{lim_min, lim_max}{Minimum and maximum limits of the plot area.}

\item{subquadrat_side}{Length in meters of the side of a subquadrat.}

\item{tag_size}{A number giving tag size. Passed to
[ggrepel::geom_text_repel].}

\item{move_edge}{A number to adjust the extension of the grid lines beyond
the plot limits.}
}
\value{
A list which each element is a plot of class ggplot.
}
\description{
Map trees within a quadrat.
}
\examples{
library(dplyr)
library(fgeo.utils)

# Fixing wrong names
vft <- rename(bciex::bci12vft_mini, QX = x, QY = y)

# Filter the data you want. For example:

# Filtering trees of diameter greater than 10 cm from the last census of plot 1
# (see also ?fgeo.utils::rm_dead_twice)
want <- dplyr::filter(
  vft,
  DBH > 10,
  CensusID == max(CensusID, na.rm = TRUE),
  PlotID == 1
)

# Filtering only two quadrats to save time
two_quadrats <- top(want, QuadratID, 2)
p <- map_quad(two_quadrats)

# Visualizing only the first plot of `p`
first(p)

# Printing all plots pf `p` to .pdf, with parameters optimized for size letter
tmp <- tempfile()  # Remplace this by somehtihing like "maps.pdf"

pdf(tmp, paper = "letter", height = 10.5, width = 8)
p
dev.off()

# Customizing the maps ----------------------------------------------------

# Filtering only one quadrat to save time
one_quad <- top(want, QuadratID)

# A custom title and header
myheader <- paste(
  " ",
  "Head column 1                     Head column 2                          ",
  " ",
  " ........................................................................",
  " ........................................................................",
  sep = "\\n"
)
map_quad(one_quad, title_quad = "My Site, 2018. Quad:", header = myheader)

# Many more tweaks are possible

# Use functions from ggplot to tweak theme
library(ggplot2)

map_quad(
  one_quad,
  title_quad = "My Site, 2018. Quad:",
  header = map_quad_header("spanish"),
  tag_size = 3,
  theme = theme_map_quad(
    axis.text = NULL,  # NULL shows axis.text; element_blank() doesn't.
    plot.title = element_text(size = 15),
    plot.subtitle = element_text(size = 5),
    panel.background = element_rect(fill = "grey")
  )
)

}
