% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tag.R
\name{map_tag}
\alias{map_tag}
\title{Title}
\usage{
map_tag(vft, dim_x = 20, div_x = 5, dim_y = dim_x, div_y = div_x,
  site_name = "Site Name, YYYY")
}
\arguments{
\item{vft}{A ViewFullTable -- a subclass of a dataframe.}
}
\description{
Title
}
