% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{paginate}
\alias{paginate}
\title{Add the variable \code{subquadrat} that really means the page the subquadrats will
print}
\usage{
paginate(x)
}
\arguments{
\item{x}{A dataframe}
}
\description{
Add the variable \code{subquadrat} that really means the page the subquadrats will
print
}
\examples{
splitted[[1]] \%>\% paginate()
}
