% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_vft.R
\name{pick_vft}
\alias{pick_vft}
\title{Pick ViewFullTables to create small example datasets.}
\usage{
pick_vft(.data, n_censuses = 1, n_quadrats = 1, n_rows = NULL)
}
\value{
A tibble (dataframe).
}
\description{
This function helps you to create small datasets for examples. It defaults to
returning the input ViewFullTable "as is", but provides arguments to pick a
specific number of censuses, quadrats and rows.
}
\examples{
# Same
dim(fgeo.x::vft_4quad)
dim(pick_vft_xxx(fgeo.x::vft_4quad))

dim(
  pick_vft(
    fgeo.x::vft_4quad, 
    n_censuses = 2, 
    n_quadrats = 2, 
    n_rows = 50
 )
)
}
\concept{datasets}
