% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_elev.R
\name{sp}
\alias{sp}
\title{Enable autoplotting the variable \code{sp}.}
\usage{
sp(sp)
}
\arguments{
\item{sp}{A ForestGEO-like dataframe with the column \code{sp}.}
}
\value{
An S3 object of class 'sp'.
}
\description{
Enable autoplotting the variable \code{sp}.
}
\examples{
class(sp(fgeo.x::stem5))

}
\seealso{
\code{\link[=autoplot.sp]{autoplot.sp()}}.

Other functions to construct fgeo classes: \code{\link{elev}},
  \code{\link{sp_elev}}

Other functions to plot species: \code{\link{autoplot.sp_elev}},
  \code{\link{autoplot_by_species.sp_elev}},
  \code{\link{sp_elev}}
}
\concept{functions to construct fgeo classes}
\concept{functions to plot species}
