% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_elev.R
\name{sp}
\alias{sp}
\title{Allow autoplotting the column \code{sp}.}
\usage{
sp(sp)
}
\arguments{
\item{sp}{A ForestGEO-like dataframe with the column \code{sp}.}
}
\value{
An S3 object of class 'sp'.
}
\description{
Allow autoplotting the column \code{sp}.
}
\examples{
assert_is_installed("fgeo.x")

inherits(sp(fgeo.x::stem5), "sp")
}
\seealso{
\code{\link[=autoplot.sp]{autoplot.sp()}}.

Other plot functions: 
\code{\link{autoplot.fgeo_habitat}()},
\code{\link{autoplot.sp_elev}()},
\code{\link{autoplot_by_species.sp_elev}()},
\code{\link{elev}()},
\code{\link{plot_dbh_bubbles_by_quadrat}()},
\code{\link{plot_tag_status_by_subquadrat}()},
\code{\link{sp_elev}()}

Other autoplots: 
\code{\link{autoplot.fgeo_habitat}()},
\code{\link{autoplot.sp_elev}()},
\code{\link{elev}()},
\code{\link{sp_elev}()}

Other functions to construct fgeo classes: 
\code{\link{elev}()},
\code{\link{sp_elev}()}

Other functions to plot species: 
\code{\link{autoplot.sp_elev}()},
\code{\link{autoplot_by_species.sp_elev}()},
\code{\link{sp_elev}()}
}
\concept{autoplots}
\concept{functions to construct fgeo classes}
\concept{functions to plot species}
\concept{plot functions}
