% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_elev.R
\name{sp_elev}
\alias{sp_elev}
\title{'Ask' \code{autoplot()} to plot the variables \code{sp} and \code{elev}.}
\usage{
sp_elev(sp, elev)
}
\arguments{
\item{sp}{A ForestGEO-like dataframe with column the column \code{sp}.}

\item{elev}{A ForestGEO-like elevation list or its \code{col} dataframe -- with
the column \code{elev}.}
}
\value{
An S3 object of class 'sp_elev'.
}
\description{
'Ask' \code{autoplot()} to plot the variables \code{sp} and \code{elev}.
}
\examples{
species_elevation <- sp_elev(
  fgeo.data::luquillo_stem5_random, fgeo.data::luquillo_elevation
)
class(species_elevation)
}
\seealso{
\code{\link[=autoplot.sp_elev]{autoplot.sp_elev()}}.

Other functions to construct fgeo classes: \code{\link{elev}},
  \code{\link{sp}}

Other functions to plot elevation: \code{\link{autoplot.elev}},
  \code{\link{autoplot.sp_elev}},
  \code{\link{autoplot_by_species.sp_elev}},
  \code{\link{elev}}

Other functions to plot species: \code{\link{autoplot.sp_elev}},
  \code{\link{autoplot.sp}},
  \code{\link{autoplot_by_species.sp_elev}},
  \code{\link{autoplot_by_species.sp}}, \code{\link{sp}}
}
\concept{functions to construct fgeo classes}
\concept{functions to plot elevation}
\concept{functions to plot species}
