% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_elev.R
\name{sp_elev}
\alias{sp_elev}
\title{Classify ForestGEO-like datasets as containing the variables \code{sp} and \code{elev}.}
\usage{
sp_elev(sp, elev)
}
\arguments{
\item{sp}{A ForestGEO-like dataframe with column the column \code{sp}.}

\item{elev}{A ForestGEO-like elevation list or its \code{col} dataframe -- with
the column \code{elev}.}
}
\value{
An S3 object of class 'sp_elev'.
}
\description{
Classify ForestGEO-like datasets as containing the variables \code{sp} and \code{elev}.
}
\examples{
species_elevation <- sp_elev(
  fgeo.data::luquillo_stem5_random, fgeo.data::luquillo_elevation
)
class(species_elevation)
}
\seealso{
\code{\link[=autoplot.sp_elev]{autoplot.sp_elev()}}.

Other functions to construct fgeo classes: \code{\link{elev}},
  \code{\link{sp}}
}
\concept{functions to construct fgeo classes}
