% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_edgy_dead.R
\name{suffix_edgy_dead}
\alias{suffix_edgy_dead}
\title{Suffix tags of dead stems that map beyond the quadrat's edge.}
\usage{
suffix_edgy_dead(x, status_d, suffix, x_q = 20, y_q = x_q)
}
\arguments{
\item{x}{A ViewFullTable.}

\item{status_d}{A character string indicaing the value of the variable 
status that corresponds to dead stems.}

\item{suffix}{A character string to suffix tree tags with.}

\item{x_q, y_q}{Integer; the x and y size in meters of a quadrat.}
}
\value{
A modified version of \code{x}.
}
\description{
Suffix tags of dead stems that map beyond the quadrat's edge.
}
\examples{
vft <- tibble::tibble(QX = 21, QY = 21, Tag = "01", Status = "dead")
vft

suffix_edgy_dead(
  x = vft, 
  status_d = "dead",
  suffix = "_d", 
  x_q = 20
) 
}
