% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_quad.R
\name{tag_dead}
\alias{tag_dead}
\title{Add the ending ".d" to the \code{Tag} of dead stems.}
\usage{
tag_dead(x, y)
}
\arguments{
\item{x}{A character vector.}

\item{y}{A character vector giving the variable Status of a fgeo-table.}
}
\value{
A character vector.
}
\description{
Add the ending ".d" to the \code{Tag} of dead stems.
}
\examples{
tag_dead(
  x = c("tag1", "tag2"),
  y = c("dead", "whatever")
)
}
