% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_dbh_bubbles}
\alias{theme_dbh_bubbles}
\title{Themes.}
\usage{
theme_dbh_bubbles(axis.text = element_blank(),
  panel.background = element_rect(fill = "white"),
  panel.grid.minor = element_blank(),
  panel.grid.major = element_blank(), legend.position = "none",
  plot.title = element_text(size = 20),
  plot.subtitle = element_text(size = 12),
  panel.border = element_rect(colour = "black", fill = NA),
  axis.ticks.length = unit(-0.1, "cm"), ...)
}
\arguments{
\item{...}{Additional arguments passed to  \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
This function outputs a theme for a specific function. You can customize any
theme by changing the default arguments and also by passing more arguments to
\code{\link[ggplot2:theme]{ggplot2::theme()}} (via \code{...}).
}
\examples{
class(theme_dbh_bubbles())

census <- fgeo.data::luquillo_tree5_random
plot_base_census(census) + theme_dbh_bubbles()
}
\seealso{
Other functions to plot dbh bubbles: \code{\link{header_dbh_bubbles}},
  \code{\link{plot_dbh_bubbles_by_quadrat}}
}
\concept{functions to plot dbh bubbles}
