% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_default}
\alias{theme_default}
\title{Themes.}
\usage{
theme_default(p, panel.grid.minor = element_line(linetype = "dashed"),
  ...)
}
\arguments{
\item{p}{A ggplot.}

\item{...}{Additional arguments passed to  \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
This function outputs a theme for a specific function. You can customize any
theme by changing the default arguments and also by passing more arguments to
\code{\link[ggplot2:theme]{ggplot2::theme()}} (via \code{...}).
}
\examples{
census <- fgeo.data::luquillo_tree5_random
theme_default(plot_base_census(census))
}
\seealso{
Other functions to create or modify plot layers: \code{\link{add_elevation_contrours}},
  \code{\link{add_elevation_labels}},
  \code{\link{add_species}}, \code{\link{axis_limits}},
  \code{\link{facet}}, \code{\link{hide}},
  \code{\link{plot_base_elevation}}
}
\concept{functions to create or modify plot layers}
