% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_default}
\alias{theme_default}
\title{Default plot theme.}
\usage{
theme_default(p, panel.grid.minor = element_line(linetype = "dashed"), ...)
}
\arguments{
\item{p}{A ggplot.}

\item{panel.grid.minor}{grid lines (\code{\link[ggplot2:element]{element_line()}}). Specify major grid lines,
or minor grid lines separately (using \code{panel.grid.major} or \code{panel.grid.minor})
or individually for each axis (using \code{panel.grid.major.x}, \code{panel.grid.minor.x},
\code{panel.grid.major.y}, \code{panel.grid.minor.y}).  Y axis grid lines are horizontal
and x axis grid lines are vertical. \verb{panel.grid.*.*} inherits from
\verb{panel.grid.*} which inherits from \code{panel.grid}, which in turn inherits
from \code{line}}

\item{...}{additional element specifications not part of base ggplot2. In general,
these should also be defined in the \verb{element tree} argument.}
}
\description{
Default plot theme.
}
\examples{
class(theme_default(ggplot2::ggplot(iris)))
}
\concept{functions to create or modify plot layers}
\keyword{internal}
