% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_default}
\alias{theme_default}
\title{Default plot theme.}
\usage{
theme_default(p, panel.grid.minor = element_line(linetype = "dashed"), ...)
}
\arguments{
\item{p}{A ggplot.}

\item{...}{additional element specifications not part of base ggplot2. In general,
these should also be defined in the \verb{element tree} argument.}
}
\description{
Default plot theme.
}
\examples{
class(theme_default(ggplot2::ggplot(iris)))
}
\concept{functions to create or modify plot layers}
\keyword{internal}
