% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\title{Themes.}
\arguments{
\item{...}{Additional arguments passed to  \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
This function outputs a theme for a specific function. You can customize any
theme by changing the default arguments and also by passing more arguments to
\code{\link[ggplot2:theme]{ggplot2::theme()}} (via \code{...}).
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\keyword{internal}
