% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.R
\docType{package}
\name{try-package}
\alias{try}
\alias{try-package}
\title{Try new things.}
\description{
For more info see README.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/try}
  \item Report bugs at \url{https://github.com/forestgeo/try/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com}

}
