% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vft_1quad}
\alias{vft_1quad}
\title{Small ViewFullTables from Luquillo.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 461 rows and 32 columns.}
\usage{
vft_1quad
}
\description{
One quadrat of a ViewFullTable from Luquillo.
}
\examples{
str(vft_1quad)
}
\seealso{
\link[fgeo.x:vft_4quad]{fgeo.x::vft_4quad}
}
\concept{datasets}
\keyword{datasets}
