% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vft_1quad.R
\name{vft_1quad}
\alias{vft_1quad}
\title{Small ViewFullTables from Luquillo.}
\usage{
vft_1quad()
}
\value{
One quadrat of a ViewFullTable dataframe.
}
\description{
One quadrat of a ViewFullTable from Luquillo.
}
\examples{
str(vft_1quad())

}
\concept{datasets}
